% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neighborhood.R
\name{adjacency}
\alias{adjacency}
\alias{adjacency,SurfaceGeometry,numeric-method}
\alias{adjacency,SurfaceGeometry,character-method}
\alias{adjacency,SurfaceGeometry,missing-method}
\title{Get Adjacency Graph}
\usage{
adjacency(x, attr, ...)

\S4method{adjacency}{SurfaceGeometry,numeric}(x, attr)

\S4method{adjacency}{SurfaceGeometry,character}(x, attr)

\S4method{adjacency}{SurfaceGeometry,missing}(x)
}
\arguments{
\item{x}{Graph structure}

\item{attr}{Character; edge attribute for weights in igraph object. If absent, weights are 0 or 1.}

\item{...}{Additional arguments}
}
\value{
A sparse adjacency matrix of class \code{dgCMatrix}
}
\description{
Get Adjacency Graph
}
