% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{adjacency}
\alias{adjacency}
\title{get adjacency graph}
\usage{
adjacency(x, attr, ...)
}
\arguments{
\item{x}{graph structure}

\item{...}{extra args}

\item{weights}{a \code{character} string indicating the edge attribute in the \code{igraph} object for the weights. If absent, weights are 0, 1.}
}
\description{
get adjacency graph
}
