% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{AFNISurfaceFileDescriptor-class}
\alias{AFNISurfaceFileDescriptor-class}
\title{AFNISurfaceFileDescriptor}
\value{
An object of class \code{AFNISurfaceFileDescriptor}.
}
\description{
This class supports the AFNI 1D file format for surface-based data
}
