% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{alpha_channel}
\alias{alpha_channel}
\title{alpha_channel}
\usage{
alpha_channel(x, ...)
}
\arguments{
\item{x}{the object to extract alpha channel from}

\item{...}{extra args}
}
\description{
extract the alpha channel
}
