% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{Arith,NeuroSurface,NeuroSurfaceVector-method}
\alias{Arith,NeuroSurface,NeuroSurfaceVector-method}
\alias{Arith,NeuroSurfaceVector,NeuroSurface-method}
\title{Arithmetic Operations between NeuroSurface and NeuroSurfaceVector}
\usage{
\S4method{Arith}{NeuroSurface,NeuroSurfaceVector}(e1, e2)

\S4method{Arith}{NeuroSurfaceVector,NeuroSurface}(e1, e2)
}
\arguments{
\item{e1}{NeuroSurface or NeuroSurfaceVector object}

\item{e2}{NeuroSurface or NeuroSurfaceVector object}
}
\value{
NeuroSurfaceVector object with arithmetic operation results
}
\description{
Arithmetic Operations between NeuroSurface and NeuroSurfaceVector
}
