% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{Arith,NeuroSurfaceVector,NeuroSurfaceVector-method}
\alias{Arith,NeuroSurfaceVector,NeuroSurfaceVector-method}
\alias{Arith,NeuroSurfaceVector,numeric-method}
\alias{Arith,numeric,NeuroSurfaceVector-method}
\alias{Arith,NeuroSurfaceVector,NeuroSurface-method}
\title{Arithmetic Operations for NeuroSurfaceVector Objects}
\usage{
\S4method{Arith}{NeuroSurfaceVector,NeuroSurfaceVector}(e1, e2)

\S4method{Arith}{NeuroSurfaceVector,numeric}(e1, e2)

\S4method{Arith}{numeric,NeuroSurfaceVector}(e1, e2)

\S4method{Arith}{NeuroSurfaceVector,NeuroSurface}(e1, e2)
}
\arguments{
\item{e1}{NeuroSurfaceVector object or numeric value}

\item{e2}{NeuroSurfaceVector object or numeric value}
}
\value{
NeuroSurfaceVector object with arithmetic operation results
}
\description{
Arithmetic Operations for NeuroSurfaceVector Objects
}
