% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{as}
\alias{as}
\title{Coercion Methods for NeuroSurface Objects}
\value{
The converted object (e.g., a numeric vector when converting to "vector").
}
\description{
Methods to convert NeuroSurface objects to other types.
}
