% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{as.vector,BrainSurface-method}
\alias{as.vector,BrainSurface-method}
\title{convert \code{BrainSurface} instance to vector}
\usage{
\S4method{as.vector}{BrainSurface}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{BrainSurface} instance to vector
}

