% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{as.vector,NeuroSurface-method}
\alias{as.vector,NeuroSurface-method}
\title{convert \code{NeuroSurface} instance to vector}
\usage{
\S4method{as.vector}{NeuroSurface}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert \code{NeuroSurface} instance to vector
}
