% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BilatBrainSurfaceVector-class}
\alias{BilatBrainSurfaceVector-class}
\title{BilatBrainSurfaceVector}
\description{
BilatBrainSurfaceVector
}
\section{Slots}{

\describe{
\item{\code{left}}{the surface data for left hemisphere, an instance of \code{BrainSurfaceVector}}

\item{\code{right}}{the surface data for right hemisphere, an instance of \code{BrainSurfaceVector}}
}}

