% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{blend_colors}
\alias{blend_colors}
\title{blend_colors}
\usage{
blend_colors(bottom, top, alpha)
}
\arguments{
\item{bottom}{the bottom color plane}

\item{top}{the top color plane}

\item{alpha}{the alpha blending value}
}
\description{
blend two color planes
}
