% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{BrainSurface}
\alias{BrainSurface}
\title{BrainSurface}
\usage{
BrainSurface(geometry, indices, data)
}
\arguments{
\item{geometry}{a \code{SurfaceGeometry} instance}

\item{indices}{an integer vector specifying the valid surface nodes.}

\item{data}{a \code{vector} of data values.}
}
\description{
construct a new BrainSurface object
}
