% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic.R
\name{clear_geodesic_cache}
\alias{clear_geodesic_cache}
\title{Clear geodesic cache}
\usage{
clear_geodesic_cache()
}
\value{
TRUE (invisibly)
}
\description{
Clear geodesic cache
}
\examples{
clear_geodesic_cache()

}
