% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{cluster_threshold}
\alias{cluster_threshold}
\title{apply a cluster-extent threshold to surface data}
\usage{
cluster_threshold(x, threshold, size, ...)
}
\arguments{
\item{x}{the object to threshold}

\item{threshold}{the numeric threshold range}

\item{size}{the minimum cluster size in nodes}

\item{...}{extra args}
}
\description{
apply a cluster-extent threshold to surface data
}
\seealso{
conn_comp
}
