% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{as_rgb}
\alias{as_rgb}
\alias{as_hexcol}
\title{color conversion}
\usage{
as_rgb(x, ...)

as_hexcol(x, ...)
}
\arguments{
\item{x}{the object to convert}

\item{...}{extra args}
}
\description{
convert to rgb colors
}
