% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{Compare,NeuroSurfaceVector,NeuroSurfaceVector-method}
\alias{Compare,NeuroSurfaceVector,NeuroSurfaceVector-method}
\title{Comparison Operations for NeuroSurfaceVector Objects}
\usage{
\S4method{Compare}{NeuroSurfaceVector,NeuroSurfaceVector}(e1, e2)
}
\arguments{
\item{e1}{NeuroSurfaceVector object or numeric value}

\item{e2}{NeuroSurfaceVector object or numeric value}
}
\value{
NeuroSurfaceVector object with comparison results
}
\description{
Comparison Operations for NeuroSurfaceVector Objects
}
