% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_hull_world_cpp}
\alias{compute_hull_world_cpp}
\title{Compute boundary hull points in world space (C++)}
\usage{
compute_hull_world_cpp(vol, dims, grid2world, thresh, n_points, mask)
}
\arguments{
\item{vol}{flattened numeric array (dim = dims)}

\item{dims}{integer vector length 3 (nx, ny, nz)}

\item{grid2world}{4x4 matrix mapping voxel indices to world coords}

\item{thresh}{intensity threshold}

\item{n_points}{max number of hull points (approximate)}

\item{mask}{optional logical vector same length as vol; if empty, ignored}
}
\value{
numeric matrix (n_points x 3) of (x,y,z) in world coords
}
\description{
Compute boundary hull points in world space (C++)
}
