% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{conn_comp,NeuroSurfaceVector-method}
\alias{conn_comp,NeuroSurfaceVector-method}
\title{connected components on a surface}
\usage{
\S4method{conn_comp}{NeuroSurfaceVector}(x, threshold, index = 1)
}
\description{
connected components on a surface
}
