% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R, R/geometry.R, R/neuro_surface.R
\name{coords,ROISurface-method}
\alias{coords,ROISurface-method}
\alias{coords,SurfaceGeometry-method}
\alias{coords,igraph-method}
\alias{coords,NeuroSurfaceVector-method}
\alias{coords,NeuroSurface-method}
\title{coords}
\usage{
\S4method{coords}{ROISurface}(x)

\S4method{coords}{SurfaceGeometry}(x)

\S4method{coords}{igraph}(x)

\S4method{coords}{NeuroSurfaceVector}(x)

\S4method{coords}{NeuroSurface}(x)
}
\arguments{
\item{x}{the object to extract cooords from}
}
\description{
extract coordinates
}
