% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curv.R
\name{curvature}
\alias{curvature}
\alias{curv}
\title{Convenience Alias for Curvature Computation}
\usage{
curvature(x, ...)
}
\arguments{
\item{x}{Object to compute curvature from}

\item{...}{Additional arguments}
}
\value{
Curvature vector
}
\description{
This is a convenience wrapper around \code{\link{curvature}}.
}
