% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{data_reader,SurfaceGeometryMetaInfo-method}
\alias{data_reader,SurfaceGeometryMetaInfo-method}
\alias{data_reader,NIMLSurfaceDataMetaInfo-method}
\title{data_reader}
\usage{
\S4method{data_reader}{SurfaceGeometryMetaInfo}(x)

\S4method{data_reader}{NIMLSurfaceDataMetaInfo}(x)
}
\arguments{
\item{x}{object used to create reader from}
}
\description{
construct a reader function
}
