% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{debug_surfwidget}
\alias{debug_surfwidget}
\title{Debugging Helper for surfwidget}
\usage{
debug_surfwidget(x, verbose = TRUE)
}
\arguments{
\item{x}{The surface object you're trying to visualize}

\item{verbose}{Logical, whether to print detailed information}
}
\value{
Invisibly returns a list of diagnostic information
}
\description{
This function helps diagnose issues with surfwidget rendering by checking
common problems and providing debugging information.
}
\examples{
\donttest{
# Load a surface and check it
surf_file <- system.file("extdata", "std.8_lh.smoothwm.asc", package="neurosurf")
surf <- read_surf(surf_file)
debug_surfwidget(surf)
}

}
