% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_plot.R
\name{.ns_autocrop}
\alias{.ns_autocrop}
\title{Auto-crop a raster image (removes whitespace)}
\usage{
.ns_autocrop(img, border = 0)
}
\description{
Auto-crop a raster image (removes whitespace)
}
\keyword{internal}
