% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{fill,BrainSurface,list-method}
\alias{fill,BrainSurface,list-method}
\alias{fill,BrainSurface,matrix-method}
\title{fill}
\usage{
\S4method{fill}{BrainSurface,list}(x, lookup)

\S4method{fill}{BrainSurface,matrix}(x, lookup)
}
\description{
fill

fill
}

