% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{FreesurferBinarySurfaceFileDescriptor-class}
\alias{FreesurferBinarySurfaceFileDescriptor-class}
\title{FresurferBinarySurfaceFileDescriptor}
\value{
An object of class \code{FreesurferBinarySurfaceFileDescriptor}.
}
\description{
This class supports the Freesurfer binary file format for surface geometry
}
\examples{
\donttest{
# Create a FreesurferBinarySurfaceFileDescriptor object
fs_binary_descriptor <- new("FreesurferBinarySurfaceFileDescriptor")

# This descriptor would typically be used when loading Freesurfer binary surfaces
# Example of a path to a Freesurfer binary surface file:
# fs_binary_file <- "/path/to/subject/surf/lh.pial"

# In practice, this descriptor would be used in code like:
# surface_geo <- loadSurfaceGeometry(fs_binary_file, descriptor = fs_binary_descriptor)
}

}
