% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{geometry}
\alias{geometry}
\alias{geometry,NeuroSurface-method}
\alias{geometry,NeuroSurfaceVector-method}
\title{Extract Geometry from Surface Object}
\usage{
geometry(x)

\S4method{geometry}{NeuroSurface}(x)

\S4method{geometry}{NeuroSurfaceVector}(x)
}
\arguments{
\item{x}{A surface object}
}
\value{
A geometry object representing the surface structure
}
\description{
Extracts the geometric representation from a surface object.
}
\seealso{
\code{\link{vertices}}, \code{\link{nodes}}
}
