% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_set.R
\name{get_surface}
\alias{get_surface}
\title{Retrieve a geometry from a SurfaceSet}
\usage{
get_surface(x, label = NULL)
}
\arguments{
\item{x}{SurfaceSet}

\item{label}{Label of the desired surface variant. Defaults to the set's default.}
}
\value{
A `SurfaceGeometry` object.
}
\description{
Retrieve a geometry from a SurfaceSet
}
\examples{
geom <- example_surface_geometry()
ss <- surface_set(inflated = geom)
get_surface(ss, "inflated")

}
