% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{GIFTISurfaceDataMetaInfo}
\alias{GIFTISurfaceDataMetaInfo}
\title{Constructor for \code{GIFTISurfaceDataMetaInfo} class}
\usage{
GIFTISurfaceDataMetaInfo(descriptor, header)
}
\arguments{
\item{descriptor}{the file descriptor}

\item{header}{a \code{list} containing header information}
}
\description{
Constructor for \code{GIFTISurfaceDataMetaInfo} class
}
