% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R, R/neuro_surface.R
\name{indices,ROISurface-method}
\alias{indices,ROISurface-method}
\alias{indices,ROISurfaceVector-method}
\alias{indices,NeuroSurfaceVector-method}
\alias{indices,NeuroSurface-method}
\title{indices}
\usage{
\S4method{indices}{ROISurface}(x)

\S4method{indices}{ROISurfaceVector}(x)

\S4method{indices}{NeuroSurfaceVector}(x)

\S4method{indices}{NeuroSurface}(x)
}
\arguments{
\item{x}{the object to extract indices from}
}
\description{
extract indices
}
