% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neighborhood.R
\name{laplacian}
\alias{laplacian}
\alias{laplacian,SurfaceGeometry,missing,missing-method}
\alias{laplacian,SurfaceGeometry,missing,numeric-method}
\title{Compute Graph Laplacian}
\usage{
laplacian(x, normalized, weights, ...)

\S4method{laplacian}{SurfaceGeometry,missing,missing}(x)

\S4method{laplacian}{SurfaceGeometry,missing,numeric}(x, weights)
}
\arguments{
\item{x}{Object to compute Laplacian from}

\item{normalized}{Logical; whether to normalize the Laplacian}

\item{weights}{Edge weights for weighted Laplacian matrix}

\item{...}{Additional arguments}
}
\value{
A sparse Laplacian matrix of class \code{dgCMatrix}
}
\description{
Compute Graph Laplacian
}
