% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{laplacian}
\alias{laplacian}
\title{compute graph laplacian}
\usage{
laplacian(x, normalized, weights, ...)
}
\arguments{
\item{x}{the object to get Laplacian from}

\item{normalized}{logical indicating whether laplcian is normalized}

\item{weights}{edge weights for weighted Laplacian matrix}

\item{...}{extra args}
}
\description{
compute graph laplacian
}
