% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{left}
\alias{left}
\alias{left,BilatNeuroSurfaceVector-method}
\title{Get Left Hemisphere}
\usage{
left(x)

\S4method{left}{BilatNeuroSurfaceVector}(x)
}
\arguments{
\item{x}{Surface object}
}
\value{
Left hemisphere of the surface
}
\description{
Get Left Hemisphere
}
\examples{
\donttest{
# Requires bilateral surface data
# lh <- left(bilat_surface)
}

}
