% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{length,ROISurface-method}
\alias{length,ROISurface-method}
\title{Get Length of Surface Object}
\usage{
\S4method{length}{ROISurface}(x)
}
\arguments{
\item{x}{the object to extract the length from}
}
\value{
An integer giving the number of vertices
}
\description{
Returns the number of vertices in a surface object.
}
