% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\docType{methods}
\name{length,ROISurface-method}
\alias{length,ROISurface-method}
\title{length}
\usage{
\S4method{length}{ROISurface}(x)
}
\arguments{
\item{x}{the object}
}
\description{
get number of elements in object
}
