% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{loadData,BrainSurfaceVectorSource-method}
\alias{loadData,BrainSurfaceVectorSource-method}
\alias{loadData,SurfaceGeometrySource-method}
\alias{loadData,BrainSurfaceSource-method}
\alias{loadData,FreesurferSurfaceGeometryMetaInfo-method}
\title{load a BrainSurfaceVector}
\usage{
\S4method{loadData}{BrainSurfaceVectorSource}(x)

\S4method{loadData}{SurfaceGeometrySource}(x)

\S4method{loadData}{BrainSurfaceSource}(x)

\S4method{loadData}{FreesurferSurfaceGeometryMetaInfo}(x)
}
\description{
load a BrainSurfaceVector

load a \code{SurfaceGeometry} instance

load a BrainSurface

load a Freesurfer surface geometry
}
