% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{loadGIFTISurface}
\alias{loadGIFTISurface}
\title{Load GIFTI surface geometry}
\usage{
loadGIFTISurface(meta_info)
}
\arguments{
\item{meta_info}{instance of type \code{GIFTISurfaceGeometryMetaInfo}}
}
\value{
a class of type \code{SurfaceGeometry}
}
\description{
Load GIFTI surface geometry
}
\details{
requires rgl library
}
\examples{
\donttest{
# Requires GIFTI surface file
# meta <- read_meta_info(GIFTISurfaceFileDescriptor(), "surface.gii")
# geom <- loadGIFTISurface(meta)
}

}
