% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{loadSurface}
\alias{loadSurface}
\title{load a surface from a surface geometry file with optional mapped surface data}
\usage{
loadSurface(surfaceName, surfaceDataName = NULL, colind = NULL)
}
\arguments{
\item{surfaceName}{the name of the file containing the surface geometry.}

\item{surfaceDataName}{the name of the file containing the values to be mapped to the surface (optional).}

\item{colind}{the columns/samples to load (optional), only if \code{surfaceDataName} is not \code{NULL}}
}
\value{
an instance of the class:
 \code{\linkS4class{SurfaceGeometry}}
 or \code{\linkS4class{BrainSurface}}
 or \code{\linkS4class{BrainSurfaceVector}}
}
\description{
load a surface from a surface geometry file with optional mapped surface data
}

