% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{loadSurfaceDataSeq}
\alias{loadSurfaceDataSeq}
\title{loadSurfaceDataSeq}
\usage{
loadSurfaceDataSeq(leftGeometry, rightGeometry, leftDataNames, rightDataNames)
}
\arguments{
\item{leftGeometry}{a \code{\linkS4class{SurfaceGeometry}} instance for the left hemisphere}

\item{rightGeometry}{a \code{\linkS4class{SurfaceGeometry}} instance for the right hemisphere}

\item{leftDataNames}{a \code{character} vector indicating names of left-hemisphere surface data files to be mapped to geometry.}

\item{rightDataNames}{a \code{character} vector indicating names of right-hemisphere surface data files to be mapped to geometry.}
}
\description{
load one or more surface datasets for both left and right hemispheres
}
