% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{loadSurfaceGeometry}
\alias{loadSurfaceGeometry}
\title{loadSurfaceGeometry}
\usage{
loadSurfaceGeometry(surfaceName)
}
\arguments{
\item{surfaceName}{the name of the file containing the surface geometry.}
}
\description{
load surface geometry from file
}

