% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{map_values,NeuroSurface,list-method}
\alias{map_values,NeuroSurface,list-method}
\alias{map_values,NeuroSurface,matrix-method}
\title{map data using a lookup table}
\usage{
\S4method{map_values}{NeuroSurface,list}(x, lookup)

\S4method{map_values}{NeuroSurface,matrix}(x, lookup)
}
\description{
map data using a lookup table
}
