% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{map_values,NeuroSurface,matrix-method}
\alias{map_values,NeuroSurface,matrix-method}
\title{Map Values for NeuroSurface with Matrix Lookup}
\usage{
\S4method{map_values}{NeuroSurface,matrix}(x, lookup)
}
\arguments{
\item{x}{a \code{NeuroSurface} object}

\item{lookup}{a \code{matrix} with two columns: the first column is the key, and the second column is the value}
}
\description{
Map Values for NeuroSurface with Matrix Lookup
}
