% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{map_values,NeuroSurface,list-method}
\alias{map_values,NeuroSurface,list-method}
\alias{map_values,NeuroSurface,matrix-method}
\title{map_values}
\usage{
\S4method{map_values}{NeuroSurface,list}(x, lookup)

\S4method{map_values}{NeuroSurface,matrix}(x, lookup)
}
\arguments{
\item{x}{the object to map over}

\item{lookup}{the lookup table}
}
\description{
map data using a key -> value lookup table
}
