% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurosurf.R
\docType{package}
\name{neurosurf-package}
\alias{neurosurf-package}
\title{neurosurf}
\description{
Data structures and IO for surface-based neuroimaging data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/neurosurf}
  \item Report bugs at \url{https://github.com/bbuchsbaum/neurosurf/issues}
}

}
\author{
\strong{Maintainer}: Bradley R Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
