% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{neurosurf_cache_dir}
\alias{neurosurf_cache_dir}
\title{Get the synthetic fixture cache directory}
\usage{
neurosurf_cache_dir()
}
\value{
Path to the versioned cache directory.
}
\description{
Get the synthetic fixture cache directory
}
\keyword{internal}
