% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{neurosurf_has_testdata}
\alias{neurosurf_has_testdata}
\title{Check if test data is available}
\usage{
neurosurf_has_testdata(file)
}
\arguments{
\item{file}{File name.}
}
\value{
Logical indicating availability.
}
\description{
Check if test data is available
}
\keyword{internal}
