% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{neurosurf_testdata_path}
\alias{neurosurf_testdata_path}
\title{Locate bundled data or synthetic fixtures}
\usage{
neurosurf_testdata_path(file)
}
\arguments{
\item{file}{File name.}
}
\value{
File path, or an empty string if it does not exist.
}
\description{
Locate bundled data or synthetic fixtures
}
\keyword{internal}
