% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroSurfaceVectorSource-class}
\alias{NeuroSurfaceVectorSource-class}
\title{NeuroSurfaceVectorSource}
\value{
An object of class \code{NeuroSurfaceVectorSource}.
}
\description{
A class that is used to produce a \code{\linkS4class{NeuroSurfaceVector}} instance
}
\section{Slots}{

\describe{
\item{\code{geometry}}{a \code{\linkS4class{SurfaceGeometry}} instance}

\item{\code{data_meta_info}}{a \code{\linkS4class{SurfaceDataMetaInfo}} instance}

\item{\code{colind}}{the column indices vector of the surface maps to be loaded}
}}

