% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NIMLSurfaceFileDescriptor-class}
\alias{NIMLSurfaceFileDescriptor-class}
\title{NIMLSurfaceFileDescriptor}
\value{
An object of class \code{NIMLSurfaceFileDescriptor}.
}
\description{
This class supports the NIML file format for surface-based data
}
