% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/geometry.R, R/neuro_surface.R
\name{nodes}
\alias{nodes}
\alias{nodes,SurfaceGeometry-method}
\alias{nodes,NeuroSurface-method}
\alias{nodes,NeuroSurfaceVector-method}
\title{Extract Surface Node Numbers}
\usage{
nodes(x)

\S4method{nodes}{SurfaceGeometry}(x)

\S4method{nodes}{NeuroSurface}(x)

\S4method{nodes}{NeuroSurfaceVector}(x)
}
\arguments{
\item{x}{An object representing a surface.}
}
\value{
A vector of node numbers.
}
\description{
Retrieves the node numbers from a surface object.
}
\seealso{
\code{\link{vertices}}
}
