% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surface.R
\docType{methods}
\name{plot,SurfaceGeometry,missing-method}
\alias{plot,SurfaceGeometry,missing-method}
\alias{plot,NeuroSurface,missing-method}
\alias{plot,LabeledNeuroSurface,missing-method}
\title{plot a surface}
\usage{
\S4method{plot}{SurfaceGeometry,missing}(x, vals = NA,
  cmap = grDevices::gray(seq(0, 1, length.out = 255)),
  vert_clrs = NULL, irange = range(vals), thresh = c(0, 0),
  alpha = 1, specular = "black", bgcol = "lightgray", ...)

\S4method{plot}{NeuroSurface,missing}(x, cmap = grDevices::gray(seq(0, 1,
  length.out = 255)), vert_clrs = NULL, irange = range(x@data, na.rm =
  TRUE), thresh = c(0, 0), alpha = 1, specular = "black",
  bgcol = "lightgray", ...)

\S4method{plot}{LabeledNeuroSurface,missing}(x, cmap = x@cols,
  vert_clrs = NULL, irange = range(x@data, na.rm = TRUE),
  thresh = c(0, 0), alpha = 1, specular = "black",
  bgcol = "lightgray", ...)
}
\arguments{
\item{x}{the surface to plot}

\item{vals}{the \code{vector} of values at each surface node in \code{x}}

\item{cmap}{a color map consisting of a vector of colors in hex format (e.g. \code{gray(n=255)})}

\item{vert_clrs}{optional vertex colors in hex format}

\item{irange}{the intensity range indicating the low and high values of the color scale.}

\item{thresh}{a 2-element vector indicating the lower and upper transparency thresholds.}

\item{alpha}{the foreground trasnparency, default is 1 (opaque).}

\item{specular}{see\code{rgl material3d}}

\item{bgcol}{a background color or vector of colors used to shade the surface.}

\item{...}{extra args to send to \code{view_surface}}
}
\description{
plot a surface
}
