% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surface.R
\name{plot.SurfaceGeometry}
\alias{plot.SurfaceGeometry}
\title{Plot method for SurfaceGeometry objects}
\usage{
\method{plot}{SurfaceGeometry}(x, y, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{SurfaceGeometry}} object.}

\item{y}{Ignored (for S3 method compatibility).}

\item{...}{Additional arguments passed to \code{\link{view_surface}}.}
}
\value{
Invisibly returns the object ID(s) from the RGL scene.
}
\description{
Plot method for SurfaceGeometry objects
}
\examples{
\donttest{
geom <- example_surface_geometry()
plot(geom)
}

}
