% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_surf_data}
\alias{read_surf_data}
\title{load surface data and link to \code{\linkS4class{SurfaceGeometry}}}
\usage{
read_surf_data(geometry, surface_data_name, colind = NULL, nodeind = NULL)
}
\arguments{
\item{geometry}{a \code{\linkS4class{SurfaceGeometry}} instance}

\item{surface_data_name}{the name of the file containing the values to be mapped to the surface.}

\item{colind}{the subset column indices of surface dataset to load (optional)}

\item{nodeind}{the subset node indices of surface dataset to include (optional)}
}
\value{
an instance of the class \code{\linkS4class{NeuroSurface}} or \code{\linkS4class{NeuroSurfaceVector}}
}
\description{
load surface data and link to \code{\linkS4class{SurfaceGeometry}}
}
\examples{
\donttest{
# Load geometry and surface data file
# geom <- read_surf_geometry("lh.white")
# surf_data <- read_surf_data(geom, "lh.thickness")
}

}
