% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{read_surf_geometry}
\alias{read_surf_geometry}
\title{Read Surface Geometry from File}
\usage{
read_surf_geometry(surface_name)
}
\arguments{
\item{surface_name}{A character string specifying the name of the file containing the surface geometry.}
}
\value{
A \code{SurfaceGeometry} object representing the loaded surface.
}
\description{
This function loads a supported surface geometry from a file and returns a SurfaceGeometry object.
}
\details{
This function supports loading surface geometries from the following file formats:
\itemize{
  \item Freesurfer ASCII (.asc)
  \item Freesurfer binary
  \item GIFTI (.gii)
}
The appropriate loader is automatically selected based on the file extension or content.
}
\examples{
\donttest{
asc_path <- system.file("extdata", "std.8_lh.inflated.asc", package = "neurosurf")
lh_surface <- read_surf_geometry(asc_path)
}


}
