% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readAFNISurfaceHeader}
\alias{readAFNISurfaceHeader}
\title{readAFNISurfaceHeader}
\usage{
readAFNISurfaceHeader(file_name)
}
\arguments{
\item{file_name}{the name of the AFNI 1D file}
}
\description{
readAFNISurfaceHeader
}
