% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R, R/read_methods.R
\name{readFreesurferAsciiGeometry}
\alias{readFreesurferAsciiGeometry}
\title{readFreesurferAsciiGeometry}
\usage{
readFreesurferAsciiGeometry(fileName)

readFreesurferAsciiGeometry(fileName)
}
\arguments{
\item{fileName}{the file}

\item{fileName}{the file}
}
\description{
readFreesurferAsciiGeometry

readFreesurferAsciiGeometry
}

