% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readFreesurferBinaryGeometry}
\alias{readFreesurferBinaryGeometry}
\title{readFreesurferBinaryGeometry}
\usage{
readFreesurferBinaryGeometry(file_name)
}
\arguments{
\item{file_name}{the file}
}
\description{
readFreesurferBinaryGeometry
}
