% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readFreesurferBinaryHeader}
\alias{readFreesurferBinaryHeader}
\title{readFreesurferBinaryHeader}
\usage{
readFreesurferBinaryHeader(file_name)
}
\arguments{
\item{file_name}{the file}
}
\description{
readFreesurferBinaryHeader
}
