% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\docType{methods}
\name{readMetaInfo,AFNISurfaceFileDescriptor-method}
\alias{readMetaInfo,AFNISurfaceFileDescriptor-method}
\alias{readMetaInfo,NIMLSurfaceFileDescriptor-method}
\alias{readMetaInfo,FreesurferAsciiSurfaceFileDescriptor-method}
\title{readMetaInfo}
\usage{
\S4method{readMetaInfo}{AFNISurfaceFileDescriptor}(x, fileName)

\S4method{readMetaInfo}{NIMLSurfaceFileDescriptor}(x, fileName)

\S4method{readMetaInfo}{FreesurferAsciiSurfaceFileDescriptor}(x, fileName)
}
\description{
readMetaInfo

readMetaInfo

readMetaInfo
}
