% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readNIMLSurfaceHeader}
\alias{readNIMLSurfaceHeader}
\title{readNIMLSurfaceHeader}
\usage{
readNIMLSurfaceHeader(file_name)
}
\arguments{
\item{file_name}{the name of the NIML file}
}
\description{
readNIMLSurfaceHeader
}
