% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{right}
\alias{right}
\alias{right,BilatNeuroSurfaceVector-method}
\title{Get Right Hemisphere}
\usage{
right(x)

\S4method{right}{BilatNeuroSurfaceVector}(x)
}
\arguments{
\item{x}{Surface object}
}
\value{
Right hemisphere of the surface
}
\description{
Get Right Hemisphere
}
