% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{series,NeuroSurfaceVector,numeric-method}
\alias{series,NeuroSurfaceVector,numeric-method}
\alias{series,NeuroSurfaceVector,integer-method}
\alias{series,NeuroSurfaceVector,ROISurface-method}
\alias{series,NeuroSurface,numeric-method}
\title{extract a series of values for a surface vector}
\usage{
\S4method{series}{NeuroSurfaceVector,numeric}(x, i)

\S4method{series}{NeuroSurfaceVector,integer}(x, i)

\S4method{series}{NeuroSurfaceVector,ROISurface}(x, i)

\S4method{series}{NeuroSurface,numeric}(x, i)
}
\arguments{
\item{x}{the object to extract series from}

\item{i}{the indices of the series set}
}
\value{
a class of type \code{Matrix}
}
\description{
extract a series of values for a surface vector
}
