% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{series_roi,NeuroSurfaceVector,numeric-method}
\alias{series_roi,NeuroSurfaceVector,numeric-method}
\alias{series_roi,NeuroSurfaceVector,ROISurface-method}
\title{Extract ROI Time Series from Surface Vector}
\usage{
\S4method{series_roi}{NeuroSurfaceVector,numeric}(x, i)

\S4method{series_roi}{NeuroSurfaceVector,ROISurface}(x, i)
}
\arguments{
\item{x}{the object to extract the series from}

\item{i}{the indices of the series to extract}
}
\value{
An \code{\linkS4class{ROISurfaceVector}} containing the extracted time series
}
\description{
Extracts time series data from a region of interest in a surface vector.
}
