% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{series_roi,NeuroSurfaceVector,numeric-method}
\alias{series_roi,NeuroSurfaceVector,numeric-method}
\alias{series_roi,NeuroSurfaceVector,ROISurface-method}
\title{series_roi}
\usage{
\S4method{series_roi}{NeuroSurfaceVector,numeric}(x, i)

\S4method{series_roi}{NeuroSurfaceVector,ROISurface}(x, i)
}
\arguments{
\item{x}{the object o extract series from}

\item{i}{the set of indices to extract}
}
\value{
a class of type \code{ROISurfaceVector}
}
\description{
series_roi
}
