% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{show,BrainSurface-method}
\alias{show,BrainSurface-method}
\title{show a \code{BrainSurface}}
\usage{
\S4method{show}{BrainSurface}(object)
}
\arguments{
\item{x}{the object}

\item{...}{extra arguments}
}
\description{
show a \code{BrainSurface}
}

