% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{show,NeuroSurface-method}
\alias{show,NeuroSurface-method}
\title{show a \code{NeuroSurface}}
\usage{
\S4method{show}{NeuroSurface}(object)
}
\arguments{
\item{x}{the object}

\item{...}{extra arguments}
}
\description{
show a \code{NeuroSurface}
}
