% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{show,NeuroSurfaceVector-method}
\alias{show,NeuroSurfaceVector-method}
\title{show a \code{NeuroSurfaceVector}}
\usage{
\S4method{show}{NeuroSurfaceVector}(object)
}
\arguments{
\item{x}{the object}

\item{...}{extra arguments}
}
\description{
show a \code{NeuroSurfaceVector}
}
