% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{[,BrainSurfaceVector,numeric,numeric,ANY-method}
\alias{[,BrainSurfaceVector,numeric,numeric,ANY-method}
\title{extractor}
\usage{
\S4method{[}{BrainSurfaceVector,numeric,numeric,ANY}(x, i, j, ...,
  drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{...}{additional args}

\item{drop}{dimension}
}
\description{
extractor
}
