% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{[,NeuroSurfaceVector,missing,missing,ANY-method}
\alias{[,NeuroSurfaceVector,missing,missing,ANY-method}
\title{Extract All Data from NeuroSurfaceVector}
\usage{
\S4method{[}{NeuroSurfaceVector,missing,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{...}{additional args}

\item{drop}{dimension}
}
\value{
A numeric matrix containing all data
}
\description{
Extracts all data from a NeuroSurfaceVector as a matrix.
}
