% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{[,NeuroSurfaceVector,numeric,missing,ANY-method}
\alias{[,NeuroSurfaceVector,numeric,missing,ANY-method}
\title{Subset NeuroSurfaceVector by Row}
\usage{
\S4method{[}{NeuroSurfaceVector,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{...}{additional args}

\item{drop}{dimension}
}
\value{
A numeric matrix or vector of extracted row data
}
\description{
Extracts rows (vertices) from a NeuroSurfaceVector.
}
