% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{[,NeuroSurfaceVector,numeric,numeric,ANY-method}
\alias{[,NeuroSurfaceVector,numeric,numeric,ANY-method}
\title{Subset NeuroSurfaceVector}
\usage{
\S4method{[}{NeuroSurfaceVector,numeric,numeric,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index (rows/vertices)}

\item{j}{second index (columns/time points)}

\item{...}{additional args}

\item{drop}{whether to drop dimensions}
}
\value{
A numeric matrix or vector of the extracted data subset
}
\description{
Extracts a subset of data from a NeuroSurfaceVector.
}
