% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{[[,NeuroSurfaceVector,numeric,ANY-method}
\alias{[[,NeuroSurfaceVector,numeric,ANY-method}
\title{extractor}
\usage{
\S4method{[[}{NeuroSurfaceVector,numeric,ANY}(x, i)
}
\arguments{
\item{x}{the object}

\item{i}{first index}
}
\description{
extractor
}
