% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{[[,NeuroSurfaceVector,numeric-method}
\alias{[[,NeuroSurfaceVector,numeric-method}
\title{Extract Data from NeuroSurfaceVector}
\usage{
\S4method{[[}{NeuroSurfaceVector,numeric}(x, i)
}
\arguments{
\item{x}{the object}

\item{i}{first index}
}
\value{
A numeric vector of data values for the specified column
}
\description{
Extracts data from a NeuroSurfaceVector using bracket notation.
}
