% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_set.R
\name{surface_labels}
\alias{surface_labels}
\title{List available surface labels}
\usage{
surface_labels(x)
}
\arguments{
\item{x}{SurfaceSet}
}
\value{
Character vector of labels
}
\description{
List available surface labels
}
