% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{SurfaceDisk}
\alias{SurfaceDisk}
\title{Create a Region on Surface}
\usage{
SurfaceDisk(surf, index, radius, max_order = NULL)
}
\arguments{
\item{surf}{a \code{SurfaceGeometry} or \code{BrainSurface} or \code{BrainSurfaceVector}}

\item{index}{the index of the central surface node. Must be a numeric
integer value within \code{1:length(V(surf@graph))}.}

\item{radius}{the size in mm of the geodesic radius. Must be a single
positive numeric value.}

\item{max_order}{maximum number of edges to traverse.
default is computed based on average edge length.}
}
\value{
An \code{ROISurfaceVector} if \code{surf} is a \code{BrainSurface} or
  \code{BrainSurfaceVector}, otherwise an \code{ROISurface} containing the
  vertices within the specified geodesic radius.
}
\description{
Creates a Region on a Surface from a radius and surface
}
\details{
The igraph associated with \code{surf} must have an edge
  attribute named \code{dist} containing numeric weights with no
  \code{NA} values.
}
