% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{SurfaceGeometrySource-class}
\alias{SurfaceGeometrySource-class}
\title{SurfaceGeometrySource Class}
\description{
The `SurfaceGeometrySource` class serves as a factory for creating
\code{\linkS4class{SurfaceGeometry}} instances. It encapsulates the
meta information required to construct a surface geometry.
}
\details{
This class is designed to facilitate the creation of \code{\linkS4class{SurfaceGeometry}}
objects by providing a standardized way to store and access the required metadata.
It acts as an intermediate step in the process of loading and constructing
surface geometries from various file formats and sources.
}
\section{Slots}{

\describe{
\item{\code{meta_info}}{An object of class \code{\linkS4class{SurfaceGeometryMetaInfo}}
containing the metadata necessary for creating a surface geometry.}
}}

\examples{
\donttest{
# Create a SurfaceGeometryMetaInfo object
meta_info <- new("SurfaceGeometryMetaInfo",
                 header_file = "surface_meta.txt",
                 data_file = "surface_data.gii",
                 file_descriptor = new("FileFormat"),
                 vertices = 40000L,
                 faces = 79998L,
                 embed_dimension = 3L,
                 label = "white",
                 hemi = "lh")

# Create a SurfaceGeometrySource object
geom_source <- new("SurfaceGeometrySource", meta_info = meta_info)

# Use geom_source to create a SurfaceGeometry object (hypothetical function)
# surface_geometry <- createSurfaceGeometry(geom_source)
}

}
\seealso{
\code{\linkS4class{SurfaceGeometry}}, \code{\linkS4class{SurfaceGeometryMetaInfo}}
}
