% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/surfwidget.R
\name{surfwidget}
\alias{surfwidget}
\alias{surfwidget,NeuroSurface-method}
\alias{surfwidget,ColorMappedNeuroSurface-method}
\alias{surfwidget,VertexColoredNeuroSurface-method}
\title{Create a Surface Widget}
\usage{
surfwidget(x, width = NULL, height = NULL, ...)

\S4method{surfwidget}{NeuroSurface}(
  x,
  width = NULL,
  height = NULL,
  cmap = rainbow(256),
  irange = range(x@data),
  thresh = c(0, 0),
  vertexColors = NULL,
  alpha = 1,
  config = list(),
  ...
)

\S4method{surfwidget}{ColorMappedNeuroSurface}(
  x,
  width = NULL,
  height = NULL,
  thresh = c(0, 0),
  vertexColors = NULL,
  alpha = 1,
  config = list(),
  ...
)

\S4method{surfwidget}{VertexColoredNeuroSurface}(x, width = NULL, height = NULL, alpha = 1, config = list(), ...)
}
\arguments{
\item{x}{An object representing a surface (e.g., SurfaceGeometry, NeuroSurface).}

\item{width}{The width of the widget (optional).}

\item{height}{The height of the widget (optional).}

\item{...}{Additional arguments for customizing the widget appearance and behavior.}
}
\value{
An HTMLWidget object representing the surface visualization.
}
\description{
This generic function creates a widget for visualizing surface data,
allowing for different implementations based on the type of surface object.
}
\details{
The surfwidget function creates an interactive widget for visualizing
surface data, such as brain surfaces. The specific implementation depends
on the class of the object provided, allowing for customized behavior
for different types of surface representations.
}
\seealso{
\code{\link{plot_js}}, \code{\link{SurfaceGeometry}}, \code{\link{NeuroSurface}}
}
