% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{updateAlpha}
\alias{updateAlpha}
\title{Update Surface Opacity}
\usage{
updateAlpha(widget, alpha)
}
\arguments{
\item{widget}{A surfwidget object as returned by \code{\link{surfwidget}}.}

\item{alpha}{Numeric opacity value between 0 (transparent) and 1 (opaque).}
}
\description{
Adjust the overall opacity of an existing surfwidget.
}
