% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{updateColorMap}
\alias{updateColorMap}
\title{Update Surface Color Map}
\usage{
updateColorMap(widget, colorMap)
}
\arguments{
\item{widget}{A surfwidget object as returned by \code{\link{surfwidget}}.}

\item{colorMap}{A vector of colors defining the new color map.}
}
\value{
The modified surfwidget object (invisibly).
}
\description{
Change the color map used by an existing surfwidget.
}
