% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{updateIRange}
\alias{updateIRange}
\title{Update Data Intensity Range}
\usage{
updateIRange(widget, min, max)
}
\arguments{
\item{widget}{A surfwidget object as returned by \code{\link{surfwidget}}.}

\item{min}{Numeric minimum value for the intensity range.}

\item{max}{Numeric maximum value for the intensity range.}
}
\description{
Modify the minimum and maximum values used for data mapping.
}
