% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{updateSurfwidgetConfig}
\alias{updateSurfwidgetConfig}
\title{Update Surfwidget Configuration}
\usage{
updateSurfwidgetConfig(session, id, config)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{id}{The ID of the surfwidget output.}

\item{config}{A list of configuration options to update. See \code{\link{surfwidget}} for details on available options.}
}
\description{
Update the configuration of an existing surfwidget.
This sends a \code{"neurosurf-surfwidget-config"} custom message to the client.
}
