% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{updateThreshold}
\alias{updateThreshold}
\title{Update Display Threshold}
\usage{
updateThreshold(widget, min, max)
}
\arguments{
\item{widget}{A surfwidget object as returned by \code{\link{surfwidget}}.}

\item{min}{Numeric lower bound of the threshold.}

\item{max}{Numeric upper bound of the threshold.}
}
\description{
Set the threshold limits for showing surface data.
}
