% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{updateVertexColors}
\alias{updateVertexColors}
\title{Update Vertex Colors}
\usage{
updateVertexColors(widget, colors)
}
\arguments{
\item{widget}{A surfwidget object as returned by \code{\link{surfwidget}}.}

\item{colors}{A vector of colors to apply to each vertex.}
}
\description{
Replace the per-vertex colors of an existing surfwidget.
}
