% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{values,ROISurface-method}
\alias{values,ROISurface-method}
\alias{values,ROISurfaceVector-method}
\alias{values,NeuroSurface-method}
\title{Extract Data Values from Surface Objects}
\usage{
\S4method{values}{ROISurface}(x, ...)

\S4method{values}{ROISurfaceVector}(x, ...)

\S4method{values}{NeuroSurface}(x, ...)
}
\arguments{
\item{x}{the object to extract the values from}

\item{...}{additional arguments}
}
\value{
A numeric vector or matrix of data values
}
\description{
Extracts the data values associated with vertices in a surface object.
}
