% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\docType{methods}
\name{vertices}
\alias{vertices}
\alias{vertices,NeuroSurfaceVector-method}
\title{extract vertices from surface}
\usage{
vertices(x, ...)

\S4method{vertices}{NeuroSurfaceVector}(x, indices)
}
\arguments{
\item{x}{the surface object}

\item{...}{extra args}

\item{indices}{a vector of indices specifying the valid surface nodes.}
}
\description{
extract vertices from surface
}
