% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\name{writeSurfaceData}
\alias{writeSurfaceData}
\title{writeSurfaceData}
\usage{
writeSurfaceData(bsurf, outstem, hemi = "")
}
\arguments{
\item{bsurf}{a class of type \code{BrainSurface} or \code{BrainSurfaceVector}}

\item{outstem}{the name of the output file, not including extension}

\item{hemi}{name of hemisphere ("lh" or "rh")}
}
\description{
writeSurfaceData
}

