% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{adjacency,SurfaceGeometry,numeric-method}
\alias{adjacency,SurfaceGeometry,numeric-method}
\alias{adjacency,SurfaceGeometry,character-method}
\alias{adjacency,SurfaceGeometry,missing-method}
\title{adjacency}
\usage{
\S4method{adjacency}{SurfaceGeometry,numeric}(x, attr)

\S4method{adjacency}{SurfaceGeometry,character}(x, attr)

\S4method{adjacency}{SurfaceGeometry,missing}(x)
}
\description{
adjacency

adjacency

adjacency
}
