% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neighborhood.R
\name{adjacency}
\alias{adjacency}
\alias{adjacency,SurfaceGeometry,numeric-method}
\alias{adjacency,SurfaceGeometry,character-method}
\alias{adjacency,SurfaceGeometry,missing-method}
\title{get adjacency graph}
\usage{
adjacency(x, attr, ...)

\S4method{adjacency}{SurfaceGeometry,numeric}(x, attr)

\S4method{adjacency}{SurfaceGeometry,character}(x, attr)

\S4method{adjacency}{SurfaceGeometry,missing}(x)
}
\arguments{
\item{x}{graph structure}

\item{attr}{a \code{character} string indicating the edge attribute in the \code{igraph} object for the weights. If absent, weights are 0, 1.}

\item{...}{extra args}
}
\description{
get adjacency graph
}
