% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{Arith,NeuroSurface,NeuroSurface-method}
\alias{Arith,NeuroSurface,NeuroSurface-method}
\alias{Arith,NeuroSurface,NeuroSurfaceVector-method}
\title{Arithmetic Operations for NeuroSurface Objects}
\usage{
\S4method{Arith}{NeuroSurface,NeuroSurface}(e1, e2)

\S4method{Arith}{NeuroSurface,NeuroSurfaceVector}(e1, e2)
}
\arguments{
\item{e1}{the left operand}

\item{e2}{the right operand}
}
\value{
NeuroSurface object with arithmetic operation results
}
\description{
Arithmetic Operations for NeuroSurface Objects
}
