% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R, R/neuro_surface.R
\name{as.matrix,ROISurfaceVector-method}
\alias{as.matrix,ROISurfaceVector-method}
\alias{as.matrix,NeuroSurfaceVector-method}
\alias{as.matrix,BilatNeuroSurfaceVector-method}
\title{as.matrix}
\usage{
\S4method{as.matrix}{ROISurfaceVector}(x)

\S4method{as.matrix}{NeuroSurfaceVector}(x)

\S4method{as.matrix}{BilatNeuroSurfaceVector}(x)
}
\arguments{
\item{x}{the object to convert to a matrix}
}
\description{
as.matrix
}
