% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R, R/ROI.R
\docType{methods}
\name{as.matrix,BrainSurfaceVector-method}
\alias{as.matrix,BrainSurfaceVector-method}
\alias{as.matrix,ROISurfaceVector-method}
\title{convert a \code{ROISurfaceVector} to an augmented matrix}
\usage{
\S4method{as.matrix}{BrainSurfaceVector}(x)

\S4method{as.matrix}{ROISurfaceVector}(x)
}
\arguments{
\item{x}{the object}
}
\description{
convert a \code{ROISurfaceVector} to an augmented matrix
}

