% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{as.vector,NeuroSurface-method}
\alias{as.vector,NeuroSurface-method}
\title{as.vector}
\usage{
\S4method{as.vector}{NeuroSurface}(x)
}
\arguments{
\item{x}{the object to convert to a vector}
}
\description{
as.vector
}
