% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BilatNeuroSurfaceVector-class}
\alias{BilatNeuroSurfaceVector-class}
\title{Bilateral NeuroSurface Vector Class}
\description{
Represents surface data for both left and right hemispheres.
}
\section{Slots}{

\describe{
\item{\code{left}}{NeuroSurfaceVector instance for the left hemisphere}

\item{\code{right}}{NeuroSurfaceVector instance for the right hemisphere}
}}

\seealso{
\code{\link{NeuroSurfaceVector}}
}
