% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BilatNeuroSurfaceVector-class}
\alias{BilatNeuroSurfaceVector-class}
\title{BilatNeuroSurfaceVector}
\description{
BilatNeuroSurfaceVector
}
\section{Slots}{

\describe{
\item{\code{left}}{the surface data for left hemisphere, an instance of \code{NeuroSurfaceVector}}

\item{\code{right}}{the surface data for right hemisphere, an instance of \code{NeuroSurfaceVector}}
}}

