% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{cluster_threshold}
\alias{cluster_threshold}
\alias{cluster_threshold,NeuroSurfaceVector-method}
\alias{cluster_threshold,NeuroSurface-method}
\title{Apply Cluster-Extent Threshold to Surface Data}
\usage{
cluster_threshold(x, threshold, size, ...)

\S4method{cluster_threshold}{NeuroSurfaceVector}(x, threshold, size = 10, index = 1)

\S4method{cluster_threshold}{NeuroSurface}(x, threshold, size = 10)
}
\arguments{
\item{x}{Object to threshold}

\item{threshold}{the two-element threshold range to use to define connected components}

\item{size}{the minimum size of the connected components to keep}

\item{...}{Additional arguments}

\item{index}{the index/column of the underlying data matrix to cluster}
}
\value{
A thresholded surface object of the same class as \code{x}
}
\description{
Apply Cluster-Extent Threshold to Surface Data
}
\seealso{
\code{\link[neuroim2]{conn_comp}}
}
