% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{cluster_threshold,NeuroSurfaceVector-method}
\alias{cluster_threshold,NeuroSurfaceVector-method}
\alias{cluster_threshold,NeuroSurface-method}
\title{threshold based on the connected components on a surface}
\usage{
\S4method{cluster_threshold}{NeuroSurfaceVector}(x, threshold, size = 10,
  index = 1)

\S4method{cluster_threshold}{NeuroSurface}(x, threshold, size = 10)
}
\description{
threshold based on the connected components on a surface
}
