% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{cluster_threshold}
\alias{cluster_threshold}
\alias{cluster_threshold,NeuroSurfaceVector-method}
\alias{cluster_threshold,NeuroSurface-method}
\title{apply a cluster-extent threshold to surface data}
\usage{
cluster_threshold(x, threshold, size, ...)

\S4method{cluster_threshold}{NeuroSurfaceVector}(x, threshold, size = 10, index = 1)

\S4method{cluster_threshold}{NeuroSurface}(x, threshold, size = 10)
}
\arguments{
\item{x}{the object to threshold}

\item{threshold}{the numeric threshold range}

\item{size}{the minimum cluster size in nodes}

\item{...}{extra args}

\item{index}{the index/column of the underlying data matrix to cluster}
}
\description{
apply a cluster-extent threshold to surface data
}
\seealso{
conn_comp
}
