% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{cluster_threshold}
\alias{cluster_threshold}
\alias{cluster_threshold,NeuroSurfaceVector-method}
\alias{cluster_threshold,NeuroSurface-method}
\title{Apply Cluster-Extent Threshold to Surface Data}
\usage{
cluster_threshold(x, threshold, size, ...)

\S4method{cluster_threshold}{NeuroSurfaceVector}(x, threshold, size = 10, index = 1)

\S4method{cluster_threshold}{NeuroSurface}(x, threshold, size = 10)
}
\arguments{
\item{x}{Object to threshold}

\item{threshold}{Numeric threshold range}

\item{size}{Minimum cluster size in nodes}

\item{...}{Additional arguments}

\item{index}{the index/column of the underlying data matrix to cluster}
}
\description{
Apply Cluster-Extent Threshold to Surface Data
}
\seealso{
\code{\link{conn_comp}}
}
