% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ColorMappedNeuroSurface-class}
\alias{ColorMappedNeuroSurface-class}
\title{ColorMappedNeuroSurface}
\description{
A three-dimensional surface consisting of a set of triangle vertices with one value per vertex,
mapped to colors using a specified colormap and range.
}
\details{
This class extends \code{NeuroSurface} by adding color mapping functionality.
The \code{cmap} slot contains a vector of hex color codes that define the colormap.
The \code{irange} slot specifies the range of data values to be mapped to the colormap.
The \code{thresh} slot defines the visibility thresholds: data values below \code{thresh[1]}
or above \code{thresh[2]} are visible, while values in between are not visible.
}
\section{Slots}{

\describe{
\item{\code{geometry}}{The surface geometry, an instance of \code{SurfaceGeometry}}

\item{\code{indices}}{An \code{integer} vector specifying the subset of valid surface nodes encoded in the \code{geometry} object}

\item{\code{data}}{The 1-D vector of data values at each vertex of the mesh}

\item{\code{cmap}}{A character vector of hex color codes representing the colormap}

\item{\code{irange}}{A numeric vector of length 2 specifying the low and high values for color mapping}

\item{\code{thresh}}{A numeric vector of length 2 specifying the low and high thresholds for visibility}
}}

\seealso{
\code{\link{view_surface}}
}
