% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{ColorMappedNeuroSurface}
\alias{ColorMappedNeuroSurface}
\title{ColorMappedNeuroSurface}
\usage{
ColorMappedNeuroSurface(geometry, indices, data, cmap, irange, thresh)
}
\arguments{
\item{geometry}{A \code{SurfaceGeometry} object representing the underlying surface structure.}

\item{indices}{An integer vector specifying the indices of valid surface nodes.}

\item{data}{A numeric vector of data values corresponding to the surface nodes.}

\item{cmap}{A character string specifying the colormap to use for mapping the data values.}

\item{irange}{A numeric vector of length 2 specifying the range of values to map.}

\item{thresh}{A numeric value specifying the threshold for the colormap.}
}
\description{
This function creates a ColorMappedNeuroSurface object, which represents a single set of data values
associated with nodes on a surface geometry, with color mapping applied.
}
