% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arith.R
\name{Compare,NeuroSurface,numeric-method}
\alias{Compare,NeuroSurface,numeric-method}
\title{Comparison Operations for NeuroSurface Objects}
\usage{
\S4method{Compare}{NeuroSurface,numeric}(e1, e2)
}
\arguments{
\item{e1}{NeuroSurface object}

\item{e2}{Numeric value}
}
\value{
NeuroSurface object with comparison results
}
\description{
Comparison Operations for NeuroSurface Objects
}
