% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{conn_comp,NeuroSurface-method}
\alias{conn_comp,NeuroSurface-method}
\title{connected components on a surface}
\usage{
\S4method{conn_comp}{NeuroSurface}(x, threshold)
}
\description{
connected components on a surface
}
