% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{conn_comp,NeuroSurfaceVector-method}
\alias{conn_comp,NeuroSurfaceVector-method}
\alias{conn_comp,NeuroSurface-method}
\title{connected components on a surface}
\usage{
\S4method{conn_comp}{NeuroSurfaceVector}(x, threshold, index = 1)

\S4method{conn_comp}{NeuroSurface}(x, threshold)
}
\arguments{
\item{x}{the object}

\item{threshold}{the two-element threshold range to use to define connected components}

\item{index}{of the data vector to find connected components on}
}
\description{
connected components on a surface
}
