% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R, R/ROI.R
\docType{methods}
\name{coords,SurfaceGeometry-method}
\alias{coords,BrainSurface-method}
\alias{coords,BrainSurfaceVector-method}
\alias{coords,ROISurface-method}
\alias{coords,SurfaceGeometry-method}
\title{coords}
\usage{
\S4method{coords}{SurfaceGeometry}(x)

\S4method{coords}{BrainSurfaceVector}(x)

\S4method{coords}{BrainSurface}(x)

\S4method{coords}{ROISurface}(x)
}
\description{
coords

coords

coords

indices
}

