% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surface.R
\name{curv_cols}
\alias{curv_cols}
\title{convert curvature vector to a set of binary colors}
\usage{
curv_cols(vals, incol = "#B3B3B3", outcol = "#404040")
}
\arguments{
\item{vals}{the curvature values}

\item{incol}{the hex color for the sulci (inward)}

\item{outcol}{the hex color for the gyri (outward)}
}
\description{
convert curvature vector to a set of binary colors
}
