% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/curv.R, R/geometry.R
\name{curvature}
\alias{curvature}
\alias{curv}
\alias{curvature,SurfaceGeometry-method}
\title{Compute Surface Curvature Vector}
\usage{
curvature(x, ...)

curvature(x, ...)

\S4method{curvature}{SurfaceGeometry}(x)
}
\arguments{
\item{x}{Object to compute curvature from}

\item{...}{Additional arguments}
}
\value{
A numeric vector of curvature values, one per vertex
}
\description{
Compute Surface Curvature Vector
}
