% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{curvature}
\alias{curvature}
\alias{curvature,SurfaceGeometry-method}
\title{compute surface curvature vector}
\usage{
curvature(x, ...)

\S4method{curvature}{SurfaceGeometry}(x)
}
\arguments{
\item{x}{the object to get curvature from}

\item{...}{extra args}
}
\description{
compute surface curvature vector
}
