% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{data_reader,SurfaceGeometryMetaInfo-method}
\alias{data_reader,SurfaceGeometryMetaInfo-method}
\alias{data_reader,NIMLSurfaceDataMetaInfo-method}
\title{Create a Column Reader for Surface Data}
\usage{
\S4method{data_reader}{SurfaceGeometryMetaInfo}(x)

\S4method{data_reader}{NIMLSurfaceDataMetaInfo}(x)
}
\arguments{
\item{x}{A SurfaceGeometryMetaInfo object}
}
\value{
A \code{ColumnReader} object for accessing surface data columns
}
\description{
Creates a column reader object for accessing surface data.
}
