% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\docType{methods}
\name{data_reader,SurfaceGeometryMetaInfo-method}
\alias{data_reader,SurfaceGeometryMetaInfo-method}
\alias{data_reader,NIMLSurfaceDataMetaInfo-method}
\title{return a reader function}
\usage{
\S4method{data_reader}{SurfaceGeometryMetaInfo}(x)

\S4method{data_reader}{NIMLSurfaceDataMetaInfo}(x)
}
\description{
return a reader function
}
