% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_plot.R
\name{.ns_autocrop}
\alias{.ns_autocrop}
\title{Auto-crop a raster image (removes whitespace)}
\usage{
.ns_autocrop(img, border = 0)
}
\arguments{
\item{img}{A raster image array}

\item{border}{Number of pixels to preserve as border}
}
\value{
The cropped image array with whitespace removed
}
\description{
Auto-crop a raster image (removes whitespace)
}
\keyword{internal}
