% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_plot.R
\name{.ns_resize_img}
\alias{.ns_resize_img}
\title{Supersample/Resize image for antialiasing (placeholder)}
\usage{
.ns_resize_img(img, scale_factor = 0.5)
}
\description{
Supersample/Resize image for antialiasing (placeholder)
}
\keyword{internal}
