% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{find_all_neighbors}
\alias{find_all_neighbors}
\title{find all neighbors in a surface mesh within a radius}
\usage{
find_all_neighbors(surf, radius, edgeWeights, nodes = NULL,
  distance_type = c("euclidean", "geodesic", "spherical"))
}
\arguments{
\item{surf}{the \code{SurfaceGeometry} object}

\item{radius}{the spatial radius to search within}

\item{edgeWeights}{the set of edgeWeights used to compute distances}

\item{nodes}{the set of nodes to find neighbors of}

\item{distance_type}{the distance metric to use}
}
\description{
find all neighbors in a surface mesh within a radius
}
