% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{FreesurferBinarySurfaceFileDescriptor-class}
\alias{FreesurferBinarySurfaceFileDescriptor-class}
\title{FresurferBinarySurfaceFileDescriptor}
\description{
This class supports the Freesurfer binary file format for surface geometry
}
