% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{FreesurferSurfaceGeometryMetaInfo-class}
\alias{FreesurferSurfaceGeometryMetaInfo-class}
\title{FreeSurferSurfaceGeometryMetaInfo}
\description{
This class contains meta information for brain surface geometry
}
