% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSurface.R
\docType{methods}
\name{geometry}
\alias{geometry}
\alias{geometry,BrainSurface-method}
\alias{geometry,BrainSurfaceVector-method}
\title{extract geometry object}
\usage{
geometry(x)

\S4method{geometry}{BrainSurface}(x)

\S4method{geometry}{BrainSurfaceVector}(x)
}
\arguments{
\item{x}{the surface object}
}
\description{
extract geometry object
}
