% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\docType{methods}
\name{geometry}
\alias{geometry}
\alias{geometry,NeuroSurface-method}
\alias{geometry,NeuroSurfaceVector-method}
\title{extract geometry object}
\usage{
geometry(x)

\S4method{geometry}{NeuroSurface}(x)

\S4method{geometry}{NeuroSurfaceVector}(x)
}
\arguments{
\item{x}{the surface object}
}
\description{
extract geometry object
}
