% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{GIFTISurfaceDataMetaInfo-class}
\alias{GIFTISurfaceDataMetaInfo-class}
\title{GIFTISurfaceDataMetaInfo}
\description{
This class contains meta information for surface-based data for the GIFTI data format
}
\section{Slots}{

\describe{
\item{\code{info}}{the underlying \code{gifti} object returned by \code{\link{readgii}}}
}}

