% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{GIFTISurfaceFileDescriptor-class}
\alias{GIFTISurfaceFileDescriptor-class}
\title{GIFTISurfaceFileDescriptor}
\description{
This class supports the GIFTI file format for surface-based data
}
