% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{GIFTISurfaceGeometryMetaInfo-class}
\alias{GIFTISurfaceGeometryMetaInfo-class}
\title{GIFTISurfaceGeometryMetaInfo}
\description{
This class contains meta information for surface-based geometry for the GIFTI data format
}
\section{Slots}{

\describe{
\item{\code{info}}{the underlying \code{gifti} object returned by \code{\link{readgii}}}
}}

