% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{graph}
\alias{graph}
\title{extract \code{igraph} object}
\usage{
graph(x, ...)
}
\arguments{
\item{x}{the object to extract the graph from}

\item{...}{extra args}
}
\description{
extract \code{igraph} object
}
