% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/geometry.R, R/neuro_surface.R
\name{graph}
\alias{graph}
\alias{graph,SurfaceGeometry-method}
\alias{graph,NeuroSurface-method}
\alias{graph,NeuroSurfaceVector-method}
\title{extract \code{igraph} object}
\usage{
graph(x, ...)

\S4method{graph}{SurfaceGeometry}(x)

\S4method{graph}{NeuroSurface}(x, ...)

\S4method{graph}{NeuroSurfaceVector}(x, ...)
}
\arguments{
\item{x}{the object to extract the graph from}

\item{...}{extra args}
}
\description{
extract \code{igraph} object
}
