% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\docType{methods}
\name{indices,ROISurface-method}
\alias{indices,ROISurface-method}
\alias{indices,ROISurfaceVector-method}
\title{indices}
\usage{
\S4method{indices}{ROISurface}(x)

\S4method{indices}{ROISurfaceVector}(x)
}
\description{
indices

indices
}
