% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{LabeledNeuroSurface-class}
\alias{LabeledNeuroSurface-class}
\title{LabeledNeuroSurface Class}
\description{
Represents a 3D surface with labeled vertices, extending NeuroSurface.
}
\section{Slots}{

\describe{
\item{\code{labels}}{Character vector of label annotations}

\item{\code{cols}}{Character vector of hex color codes for labels}
}}

\seealso{
\code{\link{NeuroSurface}}
}
