% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{LabeledNeuroSurface-class}
\alias{LabeledNeuroSurface-class}
\title{LabeledNeuroSurface}
\description{
a three-dimensional surface consisting of a set of triangle vertices with one value per vertex and associated labels and colors.
}
\section{Slots}{

\describe{
\item{\code{labels}}{the list of label annotations}

\item{\code{cols}}{the vector of colors specified as a hex character vector}
}}

