% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{laplacian,SurfaceGeometry,missing,missing-method}
\alias{laplacian,SurfaceGeometry,missing,missing-method}
\alias{laplacian,SurfaceGeometry,missing,numeric-method}
\title{laplacian}
\usage{
\S4method{laplacian}{SurfaceGeometry,missing,missing}(x)

\S4method{laplacian}{SurfaceGeometry,missing,numeric}(x, weights)
}
\description{
laplacian

laplacian
}
