% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neighborhood.R
\docType{methods}
\name{laplacian}
\alias{laplacian}
\alias{laplacian,SurfaceGeometry,missing,missing-method}
\alias{laplacian,SurfaceGeometry,missing,numeric-method}
\title{compute graph laplacian}
\usage{
laplacian(x, normalized, weights, ...)

\S4method{laplacian}{SurfaceGeometry,missing,missing}(x)

\S4method{laplacian}{SurfaceGeometry,missing,numeric}(x, weights)
}
\arguments{
\item{x}{the object to get Laplacian from}

\item{normalized}{logical indicating whether laplcian is normalized}

\item{weights}{edge weights for weighted Laplacian matrix}

\item{...}{extra args}
}
\description{
compute graph laplacian
}
