% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{left,BilatNeuroSurfaceVector-method}
\alias{left,BilatNeuroSurfaceVector-method}
\alias{right,BilatNeuroSurfaceVector-method}
\title{left}
\usage{
\S4method{left}{BilatNeuroSurfaceVector}(x)

\S4method{right}{BilatNeuroSurfaceVector}(x)
}
\description{
left

right
}
