% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{left}
\alias{left}
\alias{left,BilatNeuroSurfaceVector-method}
\title{get left hemisphere}
\usage{
left(x)

\S4method{left}{BilatNeuroSurfaceVector}(x)
}
\arguments{
\item{x}{the surface}
}
\description{
get left hemisphere
}
