% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\docType{methods}
\name{length,ROISurface-method}
\alias{length,ROISurface-method}
\title{indices}
\usage{
\S4method{length}{ROISurface}(x)
}
\arguments{
\item{x}{the object to get \code{length}}
}
\description{
indices
}
