% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R, R/geometry.R
\name{load_data,NeuroSurfaceVectorSource-method}
\alias{load_data,NeuroSurfaceVectorSource-method}
\alias{load_data,NeuroSurfaceSource-method}
\alias{load_data,FreesurferSurfaceGeometryMetaInfo-method}
\alias{load_data,GIFTISurfaceGeometryMetaInfo-method}
\alias{load_data,SurfaceGeometrySource-method}
\title{load_data}
\usage{
\S4method{load_data}{NeuroSurfaceVectorSource}(x)

\S4method{load_data}{NeuroSurfaceSource}(x)

\S4method{load_data}{FreesurferSurfaceGeometryMetaInfo}(x)

\S4method{load_data}{GIFTISurfaceGeometryMetaInfo}(x)

\S4method{load_data}{SurfaceGeometrySource}(x)
}
\arguments{
\item{x}{the object to load data from}
}
\value{
A \code{SurfaceGeometry} object containing the loaded mesh data
}
\description{
Loads surface geometry data from a source object.
}
