% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{load_data,NeuroSurfaceVectorSource-method}
\alias{load_data,NeuroSurfaceVectorSource-method}
\alias{load_data,SurfaceGeometrySource-method}
\alias{load_data,NeuroSurfaceSource-method}
\alias{load_data,FreesurferSurfaceGeometryMetaInfo-method}
\title{load a NeuroSurfaceVector}
\usage{
\S4method{load_data}{NeuroSurfaceVectorSource}(x)

\S4method{load_data}{SurfaceGeometrySource}(x)

\S4method{load_data}{NeuroSurfaceSource}(x)

\S4method{load_data}{FreesurferSurfaceGeometryMetaInfo}(x)
}
\description{
load a NeuroSurfaceVector

load a \code{SurfaceGeometry} instance

load a NeuroSurface

load a Freesurfer surface geometry
}
