% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{load_data,NeuroSurfaceVectorSource-method}
\alias{load_data,NeuroSurfaceVectorSource-method}
\alias{load_data,SurfaceGeometrySource-method}
\alias{load_data,NeuroSurfaceSource-method}
\alias{load_data,FreesurferSurfaceGeometryMetaInfo-method}
\title{load a NeuroSurfaceVector}
\usage{
\S4method{load_data}{NeuroSurfaceVectorSource}(x)

\S4method{load_data}{SurfaceGeometrySource}(x)

\S4method{load_data}{NeuroSurfaceSource}(x)

\S4method{load_data}{FreesurferSurfaceGeometryMetaInfo}(x)
}
\arguments{
\item{x}{the data source to load}
}
\description{
load a NeuroSurfaceVector
}
