% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{loadFSSurface}
\alias{loadFSSurface}
\title{load Freesurfer ascii surface}
\usage{
loadFSSurface(meta_info, surf_to_world = diag(4))
}
\arguments{
\item{meta_info}{instance of type \code{FreesurferSurfaceGeometryMetaInfo}}

\item{surf_to_world}{Optional 4x4 affine transformation matrix from surface (tkr-RAS)
coordinates to world (scanner RAS) coordinates. Defaults to identity. For proper
alignment with volumes, this transform can be computed from FreeSurfer's vox2ras-tkr
matrix (available in orig.mgz or via mri_info --vox2ras-tkr).}
}
\value{
a class of type \code{SurfaceGeometry}
}
\description{
load Freesurfer ascii surface
}
\details{
requires rgl library
}
\examples{
\donttest{
# Requires FreeSurfer surface file
# meta <- read_meta_info(FreesurferAsciiSurfaceFileDescriptor(), "lh.pial.asc")
# geom <- loadFSSurface(meta)
}

}
