% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{loadFSSurface}
\alias{loadFSSurface}
\title{load Freesurfer ascii surface}
\usage{
loadFSSurface(meta_info)
}
\arguments{
\item{meta_info}{instance of type \code{FreesurferSurfaceGeometryMetaInfo}}
}
\value{
a class of type \code{NeuroSurface}
}
\description{
load Freesurfer ascii surface
}
\details{
requires rgl library
}
