% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{map_values,NeuroSurface,list-method}
\alias{map_values,NeuroSurface,list-method}
\title{Map Values for NeuroSurface with List Lookup}
\usage{
\S4method{map_values}{NeuroSurface,list}(x, lookup)
}
\arguments{
\item{x}{a \code{NeuroSurface} object}

\item{lookup}{a list of values to map}
}
\description{
Map Values for NeuroSurface with List Lookup
}
