% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neighborhood.R
\docType{methods}
\name{neighbor_graph}
\alias{neighbor_graph}
\alias{neighbor_graph,igraph,numeric,missing,missing-method}
\alias{neighbor_graph,igraph,numeric,missing,missing}
\alias{neighbor_graph,SurfaceGeometry,numeric,missing,missing-method}
\alias{neighbor_graph,SurfaceGeometry,numeric,numeric,missing-method}
\alias{neighbor_graph,igraph,numeric,numeric,missing}
\alias{neighbor_graph,SurfaceGeometry,numeric,numeric,integer-method}
\alias{neighbor_graph,igraph,numeric,numeric,integer}
\alias{neighbor_graph,SurfaceGeometry,numeric,missing,integer-method}
\alias{neighbor_graph,igraph,numeric,missing,integer}
\title{construct neighborhood graph from surface mesh using edge weights.}
\usage{
neighbor_graph(x, radius, edgeWeights, nodes, ...)

\S4method{neighbor_graph}{igraph,numeric,missing,missing}(x, radius,
  distance_type = c("geodesic", "euclidean", "spherical"))

\S4method{neighbor_graph}{SurfaceGeometry,numeric,missing,missing}(x,
  radius, distance_type = c("geodesic", "euclidean", "spherical"))

\S4method{neighbor_graph}{SurfaceGeometry,numeric,numeric,missing}(x,
  radius, edgeWeights, distance_type = c("geodesic", "euclidean",
  "spherical"))

\S4method{neighbor_graph}{SurfaceGeometry,numeric,numeric,integer}(x,
  radius, edgeWeights, nodes, distance_type = c("geodesic", "euclidean",
  "spherical"))

\S4method{neighbor_graph}{SurfaceGeometry,numeric,missing,integer}(x,
  radius, nodes, distance_type = c("geodesic", "euclidean", "spherical"))
}
\arguments{
\item{x}{surface mesh}

\item{radius}{the edge radius defining the neighborhood}

\item{edgeWeights}{vector of edgeWeights used to define edge distance.}

\item{nodes}{the subset of nodes to use}

\item{...}{extra arguments}

\item{distance_type}{the distance function}
}
\description{
construct neighborhood graph from surface mesh using edge weights.
}
