% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurosurf.R
\docType{package}
\name{neurosurf}
\alias{neurosurf}
\title{neurosurf: Data structures and IO for surface-based neuroimaging data}
\description{
Basic utilties for reading, writing and manipulating neuroimaging organized as a trinagle mesh.
}
\details{
none
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/neurosurf}
  \item Report bugs at \url{https://github.com/bbuchsbaum/neurosurf/issues}
}

}
\author{
\strong{Maintainer}: Bradley R Buchsbaum \email{brad.buchsbaum@gmail.com}

}
