% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{neurosurf_has_testdata}
\alias{neurosurf_has_testdata}
\title{Check if optional test data is available}
\usage{
neurosurf_has_testdata(file)
}
\arguments{
\item{file}{Name of the test data file to check.}
}
\value{
Logical indicating whether the file is available.
}
\description{
Check if optional test data is available
}
\keyword{internal}
