% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{neurosurf_testdata_path}
\alias{neurosurf_testdata_path}
\title{Get path to test data file}
\usage{
neurosurf_testdata_path(file)
}
\arguments{
\item{file}{Name of the test data file.}
}
\value{
Path to the file, or empty string if not found.
}
\description{
Returns the path to a test data file, checking both the package extdata
directory and the user cache directory.
}
\keyword{internal}
