% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroSurface-class}
\alias{NeuroSurface-class}
\title{NeuroSurface}
\description{
a three-dimensional surface consisting of a set of triangle vertices with one value per vertex.
}
\section{Slots}{

\describe{
\item{\code{geometry}}{the surface geometry, an instance of \code{SurfaceGeometry}}

\item{\code{indices}}{an \code{integer} vector specifying the subset of valid surface nodes encoded in the \code{geometry} object.}

\item{\code{data}}{the 1-D vector of data value at each vertex of the mesh}
}}

