% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{NeuroSurface}
\alias{NeuroSurface}
\title{NeuroSurface}
\usage{
NeuroSurface(geometry, indices, data)
}
\arguments{
\item{geometry}{a \code{SurfaceGeometry} instance}

\item{indices}{an integer vector specifying the valid surface nodes.}

\item{data}{a \code{vector} of data values.}
}
\description{
construct a new NeuroSurface object
}
