% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/neuro_surface.R
\docType{class}
\name{NeuroSurfaceSource-class}
\alias{NeuroSurfaceSource-class}
\alias{NeuroSurfaceSource}
\title{NeuroSurfaceSource}
\usage{
NeuroSurfaceSource(surface_geom, surface_data_name, colind = NULL,
  nodeind = NULL)
}
\arguments{
\item{surface_geom}{the name of the file containing the surface geometry or a \code{SurfaceGeometry} instance}

\item{surface_data_name}{the name of the file containing the data values to be mapped to the surface.}

\item{colind}{the subset of column indices to load from surface data matrix (if provided)}

\item{nodeind}{the subset of node indices to load from surface data matrix (if provided)}
}
\description{
A class that is used to produce a \code{\linkS4class{NeuroSurface}} instance
}
\section{Slots}{

\describe{
\item{\code{geometry}}{a \code{\linkS4class{SurfaceGeometry}} instance}

\item{\code{dataMetaInfo}}{a \code{\linkS4class{SurfaceDataMetaInfo}} instance}

\item{\code{colind}}{the column index of the surface map to be loaded.}

\item{\code{nodeind}}{the node indices of the surface map to be loaded.}
}}

