% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroSurfaceVector-class}
\alias{NeuroSurfaceVector-class}
\title{NeuroSurfaceVector Class}
\description{
Represents a 3D surface with multiple values per vertex.
}
\section{Slots}{

\describe{
\item{\code{geometry}}{SurfaceGeometry instance representing the surface structure}

\item{\code{indices}}{Integer vector of valid surface node indices}

\item{\code{data}}{Matrix of values, where columns represent different measures and rows correspond to surface nodes}
}}

\note{
The number of rows in 'data' must match the number of nodes in 'geometry'.
}
\seealso{
\code{\link{SurfaceGeometry}}
}
