% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{NeuroSurfaceVector-class}
\alias{NeuroSurfaceVector-class}
\title{NeuroSurfaceVector}
\description{
a three-dimensional surface consisting of a set of triangle vertices with multiple values per vertex.
}
\section{Slots}{

\describe{
\item{\code{geometry}}{the surface geometry, an instance of \code{SurfaceGeometry}}

\item{\code{indices}}{an \code{integer} vector specifying the subset of valid surface nodes from in the \code{geometry} object.}

\item{\code{data}}{a \code{Matrix} of values where each column contains a vector of values
over all the surface nodes. The number of rows of this \code{Matrix} must equal the number of nodes in the \code{geometry} object.}
}}

