% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{NeuroSurfaceVector}
\alias{NeuroSurfaceVector}
\title{NeuroSurfaceVector}
\usage{
NeuroSurfaceVector(geometry, indices, mat)
}
\arguments{
\item{geometry}{a \code{SurfaceGeometry} or \code{SurfaceSet} instance}

\item{indices}{an integer vector specifying the valid surface nodes.}

\item{mat}{a \code{matrix} of data values (rows=nodes, columns=variables)}
}
\value{
A \code{\linkS4class{NeuroSurfaceVector}} object containing the geometry, node indices, and data matrix.
}
\description{
construct a new NeuroSurfaceVector
}
