% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{NeuroSurfaceVector}
\alias{NeuroSurfaceVector}
\title{NeuroSurfaceVector}
\usage{
NeuroSurfaceVector(geometry, indices, mat)
}
\arguments{
\item{geometry}{a \code{SurfaceGeometry} instance}

\item{indices}{an integer vector specifying the valid surface nodes.}

\item{mat}{a \code{matrix} of data values (rows=nodes, columns=variables)}
}
\description{
construct a new NeuroSurfaceVector
}
