% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSurface.R
\docType{methods}
\name{nodes}
\alias{nodes}
\alias{nodes,SurfaceGeometry-method}
\alias{nodes,BrainSurface-method}
\title{extract surface node numbers}
\usage{
nodes(x)

\S4method{nodes}{SurfaceGeometry}(x)

\S4method{nodes}{BrainSurface}(x)
}
\arguments{
\item{x}{the surface object}
}
\description{
extract surface node numbers
}
