% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\name{nodes}
\alias{nodes}
\alias{nodes,SurfaceGeometry-method}
\alias{nodes,NeuroSurface-method}
\alias{nodes,NeuroSurfaceVector-method}
\title{extract surface node numbers}
\usage{
nodes(x)

\S4method{nodes}{SurfaceGeometry}(x)

\S4method{nodes}{NeuroSurface}(x)

\S4method{nodes}{NeuroSurfaceVector}(x)
}
\arguments{
\item{x}{the surface object}
}
\description{
extract surface node numbers
}
