% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surface.R
\docType{methods}
\name{plot,SurfaceGeometry-method}
\alias{plot,SurfaceGeometry-method}
\title{plot an image}
\usage{
\S4method{plot}{SurfaceGeometry}(x, vals = NA, cmap = gray(seq(0, 1,
  length.out = 255)), irange = range(vals), thresh = c(0, 0), alpha = 1,
  specular = "black", bgcol = "lightgray", ...)
}
\arguments{
\item{x}{the surface to display}

\item{vals}{the \code{vector} of values at each surface node.}

\item{cmap}{a color map consisting of a vector of colors in hex format (e.g. \code{gray(n=255)})}

\item{irange}{the intensity range indicating the low and high values of the color scale.}

\item{thresh}{a 2-element vector indicating the lower and upper transparency thresholds.}

\item{alpha}{the foreground trasnparency, default is 1 (opaque).}

\item{bgvol}{a background color or vector of colors used to shade the surface.}
}
\description{
plot an image
}
