% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/surfwidget.R
\name{plot_js}
\alias{plot_js}
\alias{plot_js,SurfaceGeometry-method}
\title{Plot Surface as an HTMLWidget}
\usage{
plot_js(x, width = NULL, height = NULL, ...)

\S4method{plot_js}{SurfaceGeometry}(x, width = NULL, height = NULL, ...)
}
\arguments{
\item{x}{Surface object to plot}

\item{width}{The width of the widget (optional)}

\item{height}{The height of the widget (optional)}

\item{...}{Additional arguments passed to the plotting function}
}
\value{
An HTMLWidget object
}
\description{
Plot Surface as an HTMLWidget
}
\examples{
\donttest{
geom <- example_surface_geometry()
widget <- plot_js(geom)
}

}
