% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searchlight.R
\name{print.Searchlight}
\alias{print.Searchlight}
\title{Print Method for Searchlight Iterator}
\usage{
\method{print}{Searchlight}(x, ...)
}
\arguments{
\item{x}{An object of class "Searchlight"}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Searchlight Iterator
}
