% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_freesurfer_annot}
\alias{read_freesurfer_annot}
\title{read a freesurfer annotation file as a \code{LabeledNeuroSurface}}
\usage{
read_freesurfer_annot(file_name, geometry)
}
\arguments{
\item{file_name}{name of the '.annot' file}

\item{geometry}{an appropriate \code{SurfaceGeometry} instance}
}
\description{
read a freesurfer annotation file as a \code{LabeledNeuroSurface}
}
