% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/IO.R
\name{read_meta_info}
\alias{read_meta_info}
\alias{read_meta_info,AFNISurfaceFileDescriptor-method}
\alias{read_meta_info,NIMLSurfaceFileDescriptor-method}
\alias{read_meta_info,FreesurferAsciiSurfaceFileDescriptor-method}
\alias{read_meta_info,FreesurferBinarySurfaceFileDescriptor-method}
\alias{read_meta_info,GIFTISurfaceFileDescriptor-method}
\title{Generic function to read image meta info given a file and a \code{\linkS4class{FileFormat}} instance.}
\usage{
read_meta_info(x, file_name)

\S4method{read_meta_info}{AFNISurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{NIMLSurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{FreesurferAsciiSurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{FreesurferBinarySurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{GIFTISurfaceFileDescriptor}(x, file_name)
}
\arguments{
\item{x}{file format}

\item{file_name}{file name contianing meta information}
}
\description{
Generic function to read image meta info given a file and a \code{\linkS4class{FileFormat}} instance.
}
