% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\docType{methods}
\name{read_meta_info,AFNISurfaceFileDescriptor-method}
\alias{read_meta_info,AFNISurfaceFileDescriptor-method}
\alias{read_meta_info,NIMLSurfaceFileDescriptor-method}
\alias{read_meta_info,FreesurferAsciiSurfaceFileDescriptor-method}
\alias{read_meta_info,FreesurferBinarySurfaceFileDescriptor-method}
\title{read_meta_info}
\usage{
\S4method{read_meta_info}{AFNISurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{NIMLSurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{FreesurferAsciiSurfaceFileDescriptor}(x,
  file_name)

\S4method{read_meta_info}{FreesurferBinarySurfaceFileDescriptor}(x,
  file_name)
}
\description{
read_meta_info

read_meta_info

read_meta_info
}
