% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_meta_info,AFNISurfaceFileDescriptor-method}
\alias{read_meta_info,AFNISurfaceFileDescriptor-method}
\alias{read_meta_info,NIMLSurfaceFileDescriptor-method}
\alias{read_meta_info,FreesurferAsciiSurfaceFileDescriptor-method}
\alias{read_meta_info,FreesurferBinarySurfaceFileDescriptor-method}
\alias{read_meta_info,GIFTISurfaceFileDescriptor-method}
\title{read_meta_info}
\usage{
\S4method{read_meta_info}{AFNISurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{NIMLSurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{FreesurferAsciiSurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{FreesurferBinarySurfaceFileDescriptor}(x, file_name)

\S4method{read_meta_info}{GIFTISurfaceFileDescriptor}(x, file_name)
}
\arguments{
\item{x}{the file descriptor object}

\item{file_name}{the name of the file containing meta infromation.}
}
\description{
read_meta_info
}
