% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{read_surf}
\alias{read_surf}
\title{read a surface data set}
\usage{
read_surf(surface_name, surface_data_name = NULL, colind = NULL,
  nodeind = NULL)
}
\arguments{
\item{surface_name}{the name of the file containing the surface geometry.}

\item{surface_data_name}{the name of the file containing the values to be mapped to the surface (optional).}

\item{colind}{the columns/samples to load (optional), only if \code{surface_data_name} is not \code{NULL}}

\item{nodeind}{the subset of node indices to load}
}
\value{
an instance of the class:
 \code{\linkS4class{SurfaceGeometry}}
 or \code{\linkS4class{NeuroSurface}}
 or \code{\linkS4class{NeuroSurfaceVector}}
}
\description{
load a surface from a surface geometry file with optional mapped surface data
}
