% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{read_surf_data_seq}
\alias{read_surf_data_seq}
\title{read_surf_data_seq}
\usage{
read_surf_data_seq(leftGeometry, rightGeometry, leftDataNames,
  rightDataNames)
}
\arguments{
\item{leftGeometry}{a \code{\linkS4class{SurfaceGeometry}} instance for the left hemisphere}

\item{rightGeometry}{a \code{\linkS4class{SurfaceGeometry}} instance for the right hemisphere}

\item{leftDataNames}{a \code{character} vector indicating names of left-hemisphere surface data files to be mapped to geometry.}

\item{rightDataNames}{a \code{character} vector indicating names of right-hemisphere surface data files to be mapped to geometry.}
}
\description{
load one or more surface datasets for both left and right hemispheres
}
