% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{read_surf_geometry}
\alias{read_surf_geometry}
\title{read_surf_geometry}
\usage{
read_surf_geometry(surface_name)
}
\arguments{
\item{surface_name}{the name of the file containing the surface geometry.}
}
\description{
load surface geometry from file
}
