% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readFreesurferAsciiGeometry}
\alias{readFreesurferAsciiGeometry}
\title{readFreesurferAsciiGeometry}
\usage{
readFreesurferAsciiGeometry(fileName)
}
\arguments{
\item{fileName}{the file}
}
\description{
readFreesurferAsciiGeometry
}
