% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{readFreesurferAsciiHeader}
\alias{readFreesurferAsciiHeader}
\title{readFreesurferAsciiHeader}
\usage{
readFreesurferAsciiHeader(fileName)
}
\arguments{
\item{fileName}{the file}
}
\description{
readFreesurferAsciiHeader
}
