% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R, R/read_methods.R
\name{readFreesurferAsciiHeader}
\alias{readFreesurferAsciiHeader}
\title{readFreesurferAsciiHeader}
\usage{
readFreesurferAsciiHeader(fileName)

readFreesurferAsciiHeader(fileName)
}
\arguments{
\item{fileName}{the file}

\item{fileName}{the file}
}
\description{
readFreesurferAsciiHeader

readFreesurferAsciiHeader
}

