% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{remeshSurface}
\alias{remeshSurface}
\title{Remesh a SurfaceGeometry object}
\usage{
remeshSurface(surfgeom, voxel_size = 2, ...)
}
\arguments{
\item{surfgeom}{A \code{SurfaceGeometry} object to be remeshed.}

\item{voxel_size}{the size of the voxels to use for the remeshing.}

\item{...}{Additional arguments to pass to \code{Rvcg::vcgUniformRemesh}.}
}
\value{
A new \code{SurfaceGeometry} object with the remeshed surface.
}
\description{
Remesh a SurfaceGeometry object
}
