% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neuro_surface.R
\docType{methods}
\name{right}
\alias{right}
\alias{right,BilatNeuroSurfaceVector-method}
\title{get right hemisphere}
\usage{
right(x)

\S4method{right}{BilatNeuroSurfaceVector}(x)
}
\arguments{
\item{x}{the surface}
}
\description{
get right hemisphere
}
