% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{ROISurface}
\alias{ROISurface}
\title{Create an instance of class \code{\linkS4class{ROISurface}}}
\usage{
ROISurface(geometry, indices, data)
}
\arguments{
\item{geometry}{the parent geometry: an instance of class \code{SurfaceGeometry}}

\item{indices}{the parent surface indices}

\item{data}{the data values, numeric \code{vector}}
}
\value{
an instance of class \code{ROISurface}
}
\description{
Create an instance of class \code{\linkS4class{ROISurface}}
}
\examples{
\donttest{
verts <- matrix(c(0,0,0,
                  1,0,0,
                  0,1,0), ncol=3, byrow=TRUE)
faces <- matrix(c(1,2,3), ncol=3, byrow=TRUE)
geom <- SurfaceGeometry(verts, faces, "lh")

ROISurface(geom, 1L, 1)

try(ROISurface(geom, 4L, 1))      # out of range
try(ROISurface(geom, 1.5, 1))     # non-integer
}
}
