% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{ROISurfaceVector}
\alias{ROISurfaceVector}
\title{Create an instance of class \code{\linkS4class{ROISurfaceVector}}}
\usage{
ROISurfaceVector(geometry, indices, data)
}
\arguments{
\item{geometry}{the parent geometry: an instance of class \code{SurfaceGeometry}}

\item{indices}{the parent surface indices}

\item{data}{the data values, a \code{matrix}}
}
\value{
an instance of class \code{ROISurfaceVector}
}
\description{
Create an instance of class \code{\linkS4class{ROISurfaceVector}}
}
\examples{
\donttest{
verts <- matrix(c(0,0,0,
                  1,0,0,
                  0,1,0), ncol=3, byrow=TRUE)
faces <- matrix(c(0L,1L,2L), ncol=3, byrow=TRUE)
geom <- SurfaceGeometry(verts, faces, "lh")

vec <- matrix(c(0.5, 1.5), nrow=1)
ROISurfaceVector(geom, c(1L,2L), vec)

try(ROISurfaceVector(geom, c(1L,4L), vec))   # out of range
try(ROISurfaceVector(geom, c(1,2.5), vec))   # non-integer
}
}
