% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{series,NeuroSurfaceVector,numeric-method}
\alias{series,NeuroSurfaceVector,numeric-method}
\alias{series,NeuroSurfaceVector,integer-method}
\alias{series,NeuroSurfaceVector,ROISurface-method}
\alias{series,NeuroSurface,numeric-method}
\title{Extract Time Series from Surface Vector}
\usage{
\S4method{series}{NeuroSurfaceVector,numeric}(x, i)

\S4method{series}{NeuroSurfaceVector,integer}(x, i)

\S4method{series}{NeuroSurfaceVector,ROISurface}(x, i)

\S4method{series}{NeuroSurface,numeric}(x, i)
}
\arguments{
\item{x}{the object to extract the series from}

\item{i}{the indices of the series to extract}
}
\value{
A matrix where columns are time points and rows are vertex indices
}
\description{
Extracts time series data from specific vertices in a surface vector.
}
