% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{series,NeuroSurfaceVector,numeric-method}
\alias{series,NeuroSurfaceVector,numeric-method}
\alias{series_roi,NeuroSurfaceVector,numeric-method}
\alias{series,NeuroSurfaceVector,integer-method}
\alias{series,NeuroSurfaceVector,ROISurface-method}
\alias{series_roi,NeuroSurfaceVector,ROISurface-method}
\alias{series,NeuroSurface,numeric-method}
\title{series}
\usage{
\S4method{series}{NeuroSurfaceVector,numeric}(x, i)

\S4method{series_roi}{NeuroSurfaceVector,numeric}(x, i)

\S4method{series}{NeuroSurfaceVector,integer}(x, i)

\S4method{series}{NeuroSurfaceVector,ROISurface}(x, i)

\S4method{series_roi}{NeuroSurfaceVector,ROISurface}(x, i)

\S4method{series}{NeuroSurface,numeric}(x, i)
}
\value{
a class of type \code{Matrix}

a class of type \code{ROISurfaceVector}
}
\description{
series

series_roi

series

series

series_roi

series
}
