% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainSurface.R
\docType{methods}
\name{series,BrainSurfaceVector,numeric-method}
\alias{series,BrainSurface,numeric-method}
\alias{series,BrainSurfaceVector,ROISurface-method}
\alias{series,BrainSurfaceVector,integer-method}
\alias{series,BrainSurfaceVector,numeric-method}
\alias{series_roi,BrainSurfaceVector,ROISurface-method}
\alias{series_roi,BrainSurfaceVector,numeric-method}
\title{series}
\usage{
\S4method{series}{BrainSurfaceVector,numeric}(x, i)

\S4method{series_roi}{BrainSurfaceVector,numeric}(x, i)

\S4method{series}{BrainSurfaceVector,integer}(x, i)

\S4method{series}{BrainSurfaceVector,ROISurface}(x, i)

\S4method{series_roi}{BrainSurfaceVector,ROISurface}(x, i)

\S4method{series}{BrainSurface,numeric}(x, i)
}
\value{
a class of type \code{Matrix}

a class of type \code{ROISurfaceVector}
}
\description{
series

series_roi

series

series

series_roi

series
}

