% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R, R/ROI.R, R/geometry.R, R/neuro_surface.R
\name{show,SurfaceGeometryMetaInfo-method}
\alias{show,SurfaceGeometryMetaInfo-method}
\alias{show,SurfaceDataMetaInfo-method}
\alias{show,ROISurface-method}
\alias{show,SurfaceGeometry-method}
\alias{show,NeuroSurfaceVector-method}
\alias{show,NeuroSurface-method}
\title{show}
\usage{
\S4method{show}{SurfaceGeometryMetaInfo}(object)

\S4method{show}{SurfaceDataMetaInfo}(object)

\S4method{show}{ROISurface}(object)

\S4method{show}{SurfaceGeometry}(object)

\S4method{show}{NeuroSurfaceVector}(object)

\S4method{show}{NeuroSurface}(object)
}
\arguments{
\item{object}{The surface object to display}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of printing.
}
\description{
Prints a human-readable summary of surface objects to the console.
}
