% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{smooth}
\alias{smooth}
\title{smooth a surface}
\usage{
smooth(x, sigma, ...)
}
\arguments{
\item{x}{the surface object to smooth}

\item{sigma}{the bandwidth}

\item{...}{extra args}
}
\description{
smooth a surface
}
