% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\docType{methods}
\name{[,NeuroSurfaceVector,missing,numeric,ANY-method}
\alias{[,NeuroSurfaceVector,missing,numeric,ANY-method}
\title{extractor}
\usage{
\S4method{[}{NeuroSurfaceVector,missing,numeric,ANY}(x, i, j, ...,
  drop = TRUE)
}
\arguments{
\item{x}{the object}

\item{i}{first index}

\item{j}{second index}

\item{...}{additional args}

\item{drop}{dimension}
}
\description{
extractor
}
