% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{SurfaceGeometry-class}
\alias{SurfaceGeometry-class}
\title{SurfaceGeometry}
\description{
a three-dimensional surface consisting of a set of triangle vertices
}
\section{Slots}{

\describe{
\item{\code{mesh}}{the underlying \code{mesh3d} object}

\item{\code{graph}}{underlying graph structure}
}}

