% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{SurfaceGeometryMetaInfo-class}
\alias{SurfaceGeometryMetaInfo-class}
\title{SurfaceGeometryMetaInfo}
\description{
This class contains meta information for brain surface geometry
}
\section{Slots}{

\describe{
\item{\code{header_file}}{name of the file containing meta information}

\item{\code{data_file}}{name of the file containing data}

\item{\code{file_descriptor}}{descriptor of image file format}

\item{\code{vertices}}{the number of surface vertices}

\item{\code{faces}}{the number of faces}

\item{\code{embed_dimension}}{the dimensionality of the embedding}

\item{\code{label}}{a label indicating the type of surface (e.g. white, pial, inflated, flat, spherical)}
}}

