% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/neuro_surface.R
\docType{class}
\name{SurfaceGeometrySource-class}
\alias{SurfaceGeometrySource-class}
\alias{SurfaceGeometrySource}
\title{SurfaceGeometrySource}
\usage{
SurfaceGeometrySource(surface_name)
}
\arguments{
\item{surface_name}{the name of the file containing the surface geometry.}
}
\description{
A class that is used to produce a \code{\linkS4class{SurfaceGeometry}} instance

Constructor for SurfaceGeometrySource
}
\section{Slots}{

\describe{
\item{\code{metaInfo}}{a \code{\linkS4class{SurfaceGeometryMetaInfo}} instance}
}}

