% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_set.R
\docType{class}
\name{SurfaceSet-class}
\alias{SurfaceSet-class}
\alias{geometry,SurfaceSet-method}
\alias{vertices,SurfaceSet-method}
\alias{faces,SurfaceSet-method}
\alias{nodes,SurfaceSet-method}
\alias{graph,SurfaceSet-method}
\alias{curvature,SurfaceSet-method}
\title{SurfaceSet: bundle multiple surface variants for one hemisphere}
\usage{
\S4method{geometry}{SurfaceSet}(x)

\S4method{vertices}{SurfaceSet}(x, ...)

\S4method{faces}{SurfaceSet}(x, ...)

\S4method{nodes}{SurfaceSet}(x)

\S4method{graph}{SurfaceSet}(x, ...)

\S4method{curvature}{SurfaceSet}(x, ...)
}
\arguments{
\item{x}{a `SurfaceSet`}

\item{...}{additional arguments forwarded to the underlying geometry method}
}
\description{
A `SurfaceSet` keeps several `SurfaceGeometry` variants (e.g., pial, white,
inflated, sphere) for the same hemisphere together with a default choice.
}
\section{Slots}{

\describe{
\item{\code{hemi}}{Character string for hemisphere (e.g., "lh" or "rh").}

\item{\code{surfaces}}{Named list of `SurfaceGeometry` objects keyed by their label.}

\item{\code{default_label}}{Character string giving the label to use by default.}
}}

