% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{surfwidgetOutput}
\alias{surfwidgetOutput}
\alias{renderSurfwidget}
\alias{updateSurfwidgetConfig}
\alias{updateIRange}
\alias{updateThreshold}
\alias{updateVertexColors}
\alias{updateAlpha}
\alias{updateZoom}
\alias{updateRotationSpeed}
\title{Shiny bindings for surfwidget}
\usage{
surfwidgetOutput(outputId, width = "100\%", height = "400px")

renderSurfwidget(expr, env = parent.frame(), quoted = FALSE)

updateSurfwidgetConfig(session, id, config)

updateIRange(widget, min, max)

updateThreshold(widget, min, max)

updateVertexColors(widget, colors)

updateAlpha(widget, alpha)

updateZoom(widget, zoom)

updateRotationSpeed(widget, speed)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a surfwidget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{id}{The ID of the surfwidget output.}

\item{config}{A list of configuration options to update. See \code{\link{surfwidget}} for details on available options.}

\item{widget}{A surfwidget htmlwidget object.}

\item{min}{Numeric lower bound of the threshold.}

\item{max}{Numeric upper bound of the threshold.}

\item{colors}{A vector of colors to apply to each vertex.}

\item{alpha}{Numeric opacity value between 0 (transparent) and 1 (opaque).}

\item{zoom}{Numeric zoom factor.}

\item{speed}{Numeric rotation speed.}
}
\value{
For \code{surfwidgetOutput}, a Shiny UI element for displaying the widget.
  For \code{renderSurfwidget}, a server-side render function.
}
\description{
Output and render functions for using surfwidget within Shiny
applications and interactive Rmd documents.

Update the configuration of an existing surfwidget.

Modify the minimum and maximum values used for data mapping.

Set the threshold limits for showing surface data.

Replace the per-vertex colors of an existing surfwidget.

Adjust the overall opacity of an existing surfwidget.

Adjust the zoom level of a surfwidget widget.

Change the automatic rotation speed of a surfwidget widget.
}
\details{
Sends a custom message of type \code{"neurosurf-surfwidget-config"} to update the widget configuration on the client.
}
\keyword{internal}
