% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfwidget.R
\name{surfwidgetOutput}
\alias{surfwidgetOutput}
\alias{renderSurfwidget}
\alias{updateSurfwidgetConfig}
\alias{updateColorMap}
\alias{updateIRange}
\alias{updateThreshold}
\alias{updateVertexColors}
\alias{updateAlpha}
\title{Shiny bindings for surfwidget}
\usage{
surfwidgetOutput(outputId, width = "100\%", height = "400px")

renderSurfwidget(expr, env = parent.frame(), quoted = FALSE)

updateSurfwidgetConfig(session, id, config)

updateColorMap(widget, colorMap)

updateIRange(widget, min, max)

updateThreshold(widget, min, max)

updateVertexColors(widget, colors)

updateAlpha(widget, alpha)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a surfwidget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{id}{The ID of the surfwidget output.}

\item{config}{A list of configuration options to update. See \code{\link{surfwidget}} for details on available options.}
}
\description{
Output and render functions for using surfwidget within Shiny
applications and interactive Rmd documents.

Update the configuration of an existing surfwidget.
}
\keyword{internal}
