% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{values,ROISurface-method}
\alias{values,ROISurface-method}
\alias{values,ROISurfaceVector-method}
\alias{values,NeuroSurface-method}
\title{values}
\usage{
\S4method{values}{ROISurface}(x, ...)

\S4method{values}{ROISurfaceVector}(x, ...)

\S4method{values}{NeuroSurface}(x, ...)
}
\arguments{
\item{x}{the object to extract the values from}

\item{...}{additional arguments}
}
\description{
values
}
