% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\docType{methods}
\name{values,ROISurface-method}
\alias{values,ROISurface-method}
\alias{values,ROISurfaceVector-method}
\title{values}
\usage{
\S4method{values}{ROISurface}(x, ...)

\S4method{values}{ROISurfaceVector}(x, ...)
}
\description{
values

values
}
