% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{VertexData-class}
\alias{VertexData-class}
\title{VertexData}
\description{
A set of arbitary vertices associated with a data table
}
\section{Slots}{

\describe{
\item{\code{indices}}{the node indices}
}}

