% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/BrainSurface.R
\docType{methods}
\name{vertices}
\alias{as.matrix,BrainSurfaceVector-method}
\alias{coords,SurfaceGeometry-method}
\alias{geometry,BrainSurface-method}
\alias{geometry,BrainSurfaceVector-method}
\alias{vertices}
\alias{vertices,BrainSurface-method}
\alias{vertices,BrainSurfaceVector-method}
\alias{vertices,SurfaceGeometry-method}
\title{extract vertices from surface}
\usage{
vertices(x, ...)

\S4method{coords}{SurfaceGeometry}(x)

\S4method{geometry}{BrainSurface}(x)

\S4method{geometry}{BrainSurfaceVector}(x)

\S4method{as.matrix}{BrainSurfaceVector}(x)

\S4method{vertices}{BrainSurface}(x)

\S4method{vertices}{BrainSurfaceVector}(x, indices)

\S4method{vertices}{SurfaceGeometry}(x, indices)
}
\arguments{
\item{x}{the surface object}

\item{...}{extra args}

\item{indices}{a vector of indices specifying the valid surface nodes.}
}
\description{
extract vertices from surface
}

