% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/geometry.R, R/neuro_surface.R
\name{vertices}
\alias{vertices}
\alias{vertices,SurfaceGeometry-method}
\alias{vertices,NeuroSurface-method}
\alias{vertices,NeuroSurfaceVector-method}
\title{Extract Vertices from a Surface Object}
\usage{
vertices(x, ...)

\S4method{vertices}{SurfaceGeometry}(x, indices)

\S4method{vertices}{NeuroSurface}(x)

\S4method{vertices}{NeuroSurfaceVector}(x, indices)
}
\arguments{
\item{x}{An object representing a surface.}

\item{...}{Additional arguments passed to methods.}

\item{indices}{a vector of indices specifying the valid surface nodes.}
}
\value{
A matrix or data structure containing vertex information.
}
\description{
Extracts the vertices from a surface object, providing a standardized 
interface across different surface representations.
}
\examples{
\dontrun{
vertex_data <- vertices(surface_obj)
num_vertices <- nrow(vertex_data)
}

}
\seealso{
\code{\link{nodes}}
}
