% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surface.R
\name{view_surface}
\alias{view_surface}
\title{display a surface via with rgl}
\usage{
view_surface(surfgeom, vals = NA, cmap = rainbow(256, alpha = 1),
  vert_clrs = NULL, bgcol = "lightgray", alpha = 1,
  add_normals = TRUE, thresh = NULL, irange = range(vals, na.rm =
  TRUE), specular = specular, viewpoint = c("lateral", "medial",
  "ventral", "posterior"), offset = c(0, 0, 0), zoom = 1, ...)
}
\arguments{
\item{surfgeom}{surface geometty of type \code{SurfaceGeometry}}

\item{vals}{the \code{vector} of values at each surface node in \code{x}}

\item{cmap}{a color map consisting of a vector of colors in hex format (e.g. \code{gray(n=255)})}

\item{vert_clrs}{optional vertex colors in hex format}

\item{bgcol}{a background color or vector of colors used to shade the surface.}

\item{alpha}{the foreground trasnparency, default is 1 (opaque).}

\item{add_normals}{whether to add_normals}

\item{thresh}{a 2-element vector indicating the lower and upper transparency thresholds.}

\item{irange}{the intensity range indicating the low and high values of the color scale.}

\item{specular}{see\code{rgl material3d}}

\item{viewpoint}{the surface viewpoint (one of: 'lateral', 'medial', 'ventral', 'posterior')}

\item{offset}{translation offset}

\item{zoom}{zoom factor}

\item{...}{args to send to rgl::shade3d}
}
\description{
display a surface via with rgl
}
