% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_to_surf.R
\name{vol_to_surf}
\alias{vol_to_surf}
\title{map values from a 3d volume to a surface in the same coordinate space}
\usage{
vol_to_surf(surf_wm, surf_pial, vol, mask = NULL, fun = c("avg", "nn",
  "mode"), knn = 6, sigma = 8, dthresh = sigma * 2)
}
\arguments{
\item{surf_wm}{the white matter (inner) surface}

\item{surf_pial}{the pial (outer) surface}

\item{vol}{an image volume of type \code{NeuroVol} that is to be mapped to surface}

\item{mask}{a mask defining the valid voxel in the image volume}

\item{fun}{the mapping function, which can be "avg" or "nn"}

\item{knn}{the number of nearest neighbors to consider (for "avg" mapping function)}

\item{sigma}{the bandwidth of the smoothing kernel (for the "avg" mapping function)}

\item{dthresh}{a valid mapping voxel only if it is less than \code{dthresh} units away from vertex.}
}
\description{
map values from a 3d volume to a surface in the same coordinate space
}
\examples{

volname <- system.file("inst/testdata/Schaefer2018_200Parcels_7Networks_order_FSLMNI152_1mm.nii", package="neurosurf")
vol <- neuroim2::read_vol(volname)
}
