% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_surface.R
\name{write_surf_data}
\alias{write_surf_data}
\title{write_surf_data}
\usage{
write_surf_data(bsurf, outstem, hemi = "")
}
\arguments{
\item{bsurf}{a class of type \code{NeuroSurface} or \code{NeuroSurfaceVector}}

\item{outstem}{the name of the output file, not including extension}

\item{hemi}{name of hemisphere ("lh" or "rh")}
}
\description{
write_surf_data
}
